% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{assert_cols_det_unique_row}
\alias{assert_cols_det_unique_row}
\title{Check a set of columns in a data frame uniquely identify
data frame rows.}
\usage{
assert_cols_det_unique_row(
  df,
  unique_key_columns,
  arg = "x",
  call = rlang::caller_env(),
  add_err_msg = ""
)
}
\arguments{
\item{df}{the dataframe to check}

\item{unique_key_columns}{Columns that, taken together, should
uniquely identify a row in the data frame.}

\item{arg}{the name of the unique grouping to check}

\item{call}{Calling environment to be passed to \code{\link[cli:cli_abort]{cli::cli_abort()}} for
traceback.}

\item{add_err_msg}{string containing an additional error message,
default is the empty string (\code{""})}
}
\value{
NULL, invisibly
}
\description{
Equivalently, this checks that when grouping by the columns in question,
each group has a single entry
}
