% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{assert_daily_data}
\alias{assert_daily_data}
\title{Assert that the vector of dates being passed in contains dates for each day}
\usage{
assert_daily_data(dates, call = rlang::caller_env(), add_err_msg = "")
}
\arguments{
\item{dates}{the vector of dates to check, must be of Date type}

\item{call}{Calling environment to be passed to \code{\link[cli:cli_abort]{cli::cli_abort()}} for
traceback.}

\item{add_err_msg}{add_err_msg string containing an additional error message,
default is the empty string (\code{""})}
}
\value{
NULL invisible
}
\description{
Assert that the vector of dates being passed in contains dates for each day
}
