% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{assert_elements_non_neg}
\alias{assert_elements_non_neg}
\title{Assert that all elements of a vector are non-negative}
\usage{
assert_elements_non_neg(
  x,
  arg = "x",
  call = rlang::caller_env(),
  add_err_msg = ""
)
}
\arguments{
\item{x}{vector of arguments to check for negativity}

\item{arg}{string to print the name of the element your checking}

\item{call}{Calling environment to be passed to \code{\link[cli:cli_abort]{cli::cli_abort()}} for
traceback.}

\item{add_err_msg}{string containing an additional error message,
default is the empty string (\code{""})}
}
\value{
NULL, invisibly
}
\description{
Assert that all elements of a vector are non-negative
}
