% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{assert_global_pop}
\alias{assert_global_pop}
\title{Assert that the vector of population sizes for the global catchment area
has only a single value}
\usage{
assert_global_pop(x, arg = "x", call = rlang::caller_env())
}
\arguments{
\item{x}{the vector to check}

\item{arg}{the name of the vector to check}

\item{call}{Calling environment to be passed to \code{\link[cli:cli_abort]{cli::cli_abort()}} for
traceback.}
}
\value{
NULL, invisibly
}
\description{
This function  checks that "global" population sizes in the data passed in as
count data contain only a single value of the catchment areas population
size. Multiple values might indicate either that there are multiple count
data streams being passed in (which is not currently supported) or that
there is a time varying population size (which is also not currently
supported). This function is specific to the
current version of the model, and will
be deprecated for a more general \code{check_for_single_value()} once
additional model functionality has been added.
}
