% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{assert_int_or_char}
\alias{assert_int_or_char}
\title{Assert that the arguments are either count type or characters}
\usage{
assert_int_or_char(x, arg = "x", call = rlang::caller_env())
}
\arguments{
\item{x}{Object with type checking}

\item{arg}{Name of the argument supplying the object}

\item{call}{Calling environment to be passed to \code{\link[cli:cli_abort]{cli::cli_abort()}}}
}
\value{
NULL, invisibly
}
\description{
This is for unique identifiers of groupings, which we will allow to
either be character or integers.
}
