% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{assert_no_dates_after_max}
\alias{assert_no_dates_after_max}
\title{Check that all dates in dataframe passed in are before a specified date}
\usage{
assert_no_dates_after_max(
  date_vector,
  max_date,
  arg_dates = "y",
  arg_max_date = "x",
  call = rlang::caller_env()
)
}
\arguments{
\item{date_vector}{vector of dates}

\item{max_date}{string indicating the maximum date in ISO8601 convention
e.g. YYYY-MM-DD}

\item{arg_dates}{string to print the name of the data you are checking the
dates for}

\item{arg_max_date}{string to print the name of the maximum date you are
checkign the data for}

\item{call}{Calling environment to be passed to \code{\link[cli:cli_abort]{cli::cli_abort()}} for
traceback.}
}
\value{
NULL, invisibly
}
\description{
This function is specifically meant to ensure that the data in the
\code{date_vector} specified does not contain dates after a given \code{max_date}. The
intended use-case for this is to ensure that one doesn't accidentally
pass in data that extends beyond the forecast date, as ideally the user
is providing vintaged "as of" datasets or at the very least is filtering the
data so that they are not including in their inference data that was
made available after the forecast was made.
}
