% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{assert_no_repeated_elements}
\alias{assert_no_repeated_elements}
\title{Check that there are no repeated elements in the vector}
\usage{
assert_no_repeated_elements(
  x,
  arg = "x",
  call = rlang::caller_env(),
  add_err_msg = ""
)
}
\arguments{
\item{x}{the vector to check}

\item{arg}{the name of the vector to check}

\item{call}{Calling environment to be passed to \code{\link[cli:cli_abort]{cli::cli_abort()}} for
traceback.}

\item{add_err_msg}{string containing an additional error message,
default is the empty string (\code{""})}
}
\value{
NULL, invisibly
}
\description{
This function  checks that the elements of a vector are
not repeated.
}
