% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{assert_no_repeats_win_group}
\alias{assert_no_repeats_win_group}
\title{Check that there are no repeated elements in groups within
a dataframe}
\usage{
assert_no_repeats_win_group(
  df,
  groups,
  arg = "x",
  call = rlang::caller_env(),
  add_err_msg = ""
)
}
\arguments{
\item{df}{the dataframe to check}

\item{groups}{a vector of strings indicating the groups of column
names that are expected to have no repeated combinations}

\item{arg}{the name of the dataframe to check}

\item{call}{Calling environment to be passed to \code{\link[cli:cli_abort]{cli::cli_abort()}} for
traceback.}

\item{add_err_msg}{string containing an additional error message,
default is the empty string (\code{""})}
}
\value{
NULL, invisibly
}
\description{
This function  checks that the elements of groups within a dataframe
dont have repeated elements
}
