% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{assert_not_empty}
\alias{assert_not_empty}
\title{Assert that the dataframe being passed to the function is not empty}
\usage{
assert_not_empty(x, arg = "x", call = rlang::caller_env(), add_err_msg = "")
}
\arguments{
\item{x}{the dataframe to check}

\item{arg}{the name of the dataframe to check}

\item{call}{Calling environment to be passed to \code{\link[cli:cli_abort]{cli::cli_abort()}} for
traceback.}

\item{add_err_msg}{add_err_msg string containing an additional error message,
default is the empty string (\code{""})}
}
\value{
NULL invisible
}
\description{
Assert that the dataframe being passed to the function is not empty
}
