% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{assert_overlap_dates}
\alias{assert_overlap_dates}
\title{Assert that a vector of dates we're testing overlaps with the comparison
dates}
\usage{
assert_overlap_dates(
  test_dates,
  comparison_dates,
  max_date,
  call = rlang::caller_env(),
  add_err_msg = ""
)
}
\arguments{
\item{test_dates}{the vector of dates to check, must be in IS08 convention}

\item{comparison_dates}{the vector of dates to compare to, must be in ISO8
convention}

\item{max_date}{the maximum date the testing dates should be}

\item{call}{Calling environment to be passed to \code{\link[cli:cli_abort]{cli::cli_abort()}} for
traceback.}

\item{add_err_msg}{add_err_msg string containing an additional error message,
default is the empty string (\code{""})}
}
\value{
NULL invisible
}
\description{
Assert that a vector of dates we're testing overlaps with the comparison
dates
}
