% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{assert_req_ww_cols_present}
\alias{assert_req_ww_cols_present}
\title{Check that the input wastewater data contains all the required column names}
\usage{
assert_req_ww_cols_present(
  ww_data,
  conc_col_name,
  lod_col_name,
  add_req_col_names = c("date", "site", "lab", "site_pop"),
  call = rlang::caller_env()
)
}
\arguments{
\item{ww_data}{tibble containing the input wastewater data}

\item{conc_col_name}{string indicating the name of the column containing
the concentration measurements in the wastewater data}

\item{lod_col_name}{string indicating the name of the column containing
the concentration measurements in the wastewater data}

\item{add_req_col_names}{vector of strings indicating the required wastewater
column names, the defaults are \code{c("date", "site", "lab", "site_pop")}}

\item{call}{Calling environment to be passed to \code{\link[cli:cli_abort]{cli::cli_abort()}} for
traceback.}
}
\value{
NULL, invisibly
}
\description{
This function is intended to be used to check that the wastewater data that
gets passed into \code{\link[=preprocess_ww_data]{preprocess_ww_data()}} contains the required columns. If
it does not, we want to tell the user which column are missing. This will not
however, ensure that the elements of the column are of the right type,
or check that the values of them make sense.
}
