% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{assert_rt_correct_length}
\alias{assert_rt_correct_length}
\title{Assert that R(t) specified for generating simulated data is of sufficient
length}
\usage{
assert_rt_correct_length(
  r_in_weeks,
  ot,
  nt,
  forecast_horizon,
  call = rlang::caller_env()
)
}
\arguments{
\item{r_in_weeks}{a vector indicating the R(t) in weeks}

\item{ot}{integer indicating the observed time: length of hospital admissions
calibration time in days}

\item{nt}{integer indicating the nowcast time: length of time between last
hospital admissions date and forecast date in days}

\item{forecast_horizon}{integer indicating the duration of the forecast in
days e.g. 28 days}

\item{call}{Calling environment to be passed to \code{\link[cli:cli_abort]{cli::cli_abort()}} for
traceback.}
}
\value{
NULL, invisible
}
\description{
Assert that R(t) specified for generating simulated data is of sufficient
length
}
