% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{assert_single_value}
\alias{assert_single_value}
\title{Assert that there is only a single value in a particular column}
\usage{
assert_single_value(x, arg = "x", call = rlang::caller_env(), add_err_msg = "")
}
\arguments{
\item{x}{the vector to check}

\item{arg}{the name of the vector to check}

\item{call}{Calling environment to be passed to \code{\link[cli:cli_abort]{cli::cli_abort()}} for
traceback.}

\item{add_err_msg}{string containing an additional error message,
default is the empty string (\code{""})}
}
\value{
NULL, invisibly
}
\description{
Assert that there is only a single value in a particular column
}
