% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{assert_site_lab_indices_align}
\alias{assert_site_lab_indices_align}
\title{Assert that the specified site and lab indices line up}
\usage{
assert_site_lab_indices_align(site, lab, call = rlang::caller_env())
}
\arguments{
\item{site}{vector of integers indicating which site (WWTP) each separate
lab-site observation comes frm}

\item{lab}{ector of integers indicating which lab the lab-site observations
come from}

\item{call}{Calling environment to be passed to \code{\link[cli:cli_abort]{cli::cli_abort()}} for
traceback.}
}
\value{
NULL, invisibly
}
\description{
Assert that the specified site and lab indices line up
}
