% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{assert_ww_site_pops_lt_total}
\alias{assert_ww_site_pops_lt_total}
\title{Assert that the sum of the wastewater site populations don't exceed
the total population}
\usage{
assert_ww_site_pops_lt_total(
  pop_size,
  ww_pop_sites,
  call = rlang::caller_env()
)
}
\arguments{
\item{pop_size}{integer indicating the population size in the hypothetical
state}

\item{ww_pop_sites}{vector indicating the population size in the
catchment area in each of those sites}

\item{call}{Calling environment to be passed to \code{\link[cli:cli_abort]{cli::cli_abort()}} for
traceback.}
}
\value{
NULL, invisibly
}
\description{
Assert that the sum of the wastewater site populations don't exceed
the total population
}
