% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_component_fwd_sim.R
\name{calc_rt}
\alias{calc_rt}
\title{Back- calculate R(t) from incident infections and the generation interval}
\usage{
calc_rt(new_i, convolve_fxn, generation_interval, uot, tot_time)
}
\arguments{
\item{new_i}{vector of numerics that spans the length of \code{tot_time},
representing the new incident infections per day}

\item{convolve_fxn}{function used to convolve infections with delay pmf}

\item{generation_interval}{vector of simplex describing the probability of
each time from infection to onwards transmission}

\item{uot}{integer indicating the days for exponential growth initialization
to occur (referred to as unobserved time)}

\item{tot_time}{integer indicating the total time we have incident
infections for}
}
\value{
a numeric vector of length(\code{tot_time} - \code{uot}) that represents
the effective reproductive number
}
\description{
Back- calculate R(t) from incident infections and the generation interval
}
\seealso{
Other Model components:
\code{\link{create_site_lab_map}()},
\code{\link{downsample_ww_obs}()},
\code{\link{format_hosp_data}()},
\code{\link{format_ww_data}()},
\code{\link{get_global_rt}()},
\code{\link{get_pred_obs_conc}()},
\code{\link{get_pred_subpop_gen_per_n}()},
\code{\link{get_time_varying_daily_ihr}()},
\code{\link{subpop_inf_process}()},
\code{\link{subpop_rt_process}()}
}
\concept{Model components}
