% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_date}
\alias{check_date}
\title{Check that all dates in dataframe passed in are before a specified date}
\usage{
check_date(df, max_date, call = rlang::caller_env())
}
\arguments{
\item{df}{dataframe with \code{date} column}

\item{max_date}{string indicating the maximum date in ISO8601 convention
e.g. YYYY-MM-DD}

\item{call}{Calling environment to be passed to the type checker}
}
\value{
NULL, invisibly
}
\description{
Check that all dates in dataframe passed in are before a specified date
}
