% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_elements_non_neg}
\alias{check_elements_non_neg}
\title{Check that all elements of a vector are non-negative}
\usage{
check_elements_non_neg(
  x,
  arg = "x",
  call = rlang::caller_env(),
  add_err_msg = ""
)
}
\arguments{
\item{x}{vector of arguments to check for negativity}

\item{arg}{string to print the name of the element your checking}

\item{call}{Calling environment to be passed to \code{\link[cli:cli_abort]{cli::cli_abort()}} for
traceback.}

\item{add_err_msg}{string containing an additional error message, defauly is
\code{NA}}
}
\value{
NULL, invisibly
}
\description{
Check that all elements of a vector are non-negative
}
