% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_for_repeat_dates}
\alias{check_for_repeat_dates}
\title{Check that there are no repeated elements in the vector of dates
corresponding to count observations}
\usage{
check_for_repeat_dates(x, arg = "x", call = rlang::caller_env())
}
\arguments{
\item{x}{the vector to check}

\item{arg}{the name of the vector to check}

\item{call}{Calling environment to be passed to \code{\link[cli:cli_abort]{cli::cli_abort()}} for
traceback.}
}
\value{
NULL, invisibly
}
\description{
This function  checks that the dates in the data passed in as count data are
not repeated, which might indicate multiple count data stream. This
functionality is not currently supported. This function is specific to the
current version of the model, and will
be deprecated for a more general \code{check_for_repeat_elements()} once
additional model functionality has been added.
}
