% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_int}
\alias{check_int}
\title{Check that the arguments are integers}
\usage{
check_int(x, arg = "x", call = rlang::caller_env())
}
\arguments{
\item{x}{Object with type checking}

\item{arg}{Name of the argument supplying the object}

\item{call}{Calling environment to be passed to \code{\link[cli:cli_abort]{cli::cli_abort()}}}
}
\value{
NULL, invisibly
}
\description{
Check that the arguments are integers
}
