% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_no_missingness}
\alias{check_no_missingness}
\title{Check there is no missignness in a particular vector}
\usage{
check_no_missingness(x, arg = "x", call = rlang::caller_env())
}
\arguments{
\item{x}{the vector to check}

\item{arg}{the name of the vector to check}

\item{call}{Calling environment to be passed to \code{\link[cli:cli_abort]{cli::cli_abort()}} for
traceback.}
}
\value{
NULL, invisibly
}
\description{
Check there is no missignness in a particular vector
}
