% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_required_hosp_inputs}
\alias{check_required_hosp_inputs}
\title{Check that the input hosp data contains all the required column names}
\usage{
check_required_hosp_inputs(
  hosp_data,
  count_col_name,
  pop_size_col_name,
  add_req_col_names = c("date"),
  call = rlang::caller_env()
)
}
\arguments{
\item{hosp_data}{tibble containing the input count data}

\item{count_col_name}{string indicating the name of the column containing
the count data}

\item{pop_size_col_name}{string indicating the name of the column containing
the population size of the count catchment area}

\item{add_req_col_names}{vector of strings indicating the required count
data column names, the defaults is \code{"date"}}

\item{call}{Calling environment to be passed to \code{\link[cli:cli_abort]{cli::cli_abort()}} for
traceback.}
}
\value{
NULL, invisibly
}
\description{
This function is intended to be used to check that the hosp data that
gets passed into \code{\link[=preprocess_hosp_data]{preprocess_hosp_data()}} contains the required columns. If
it does not, we want to tell the user which columns are missing. This will
not however, ensure that the elements of the column are of the right type,
or check that the values of them make sense.
}
