% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_required_ww_inputs}
\alias{check_required_ww_inputs}
\title{Check that the input wastewater data contains all the required column names}
\usage{
check_required_ww_inputs(
  ww_data,
  conc_col_name,
  lod_col_name,
  call = rlang::caller_env()
)
}
\arguments{
\item{ww_data}{tibble containing the input wastewater data}

\item{conc_col_name}{string indicating the name of the column containing
the concentration measurements in the wastewater data}

\item{lod_col_name}{string indicating the name of the column containing
the concentration measurements in the wastewater data}

\item{call}{Calling environment to be passed to \code{\link[cli:cli_abort]{cli::cli_abort()}} for
traceback.}
}
\value{
NULL, invisibly
}
\description{
This function is intended to be used to check that the wastewater data that
gets passed into \code{\link[=preprocess_ww_data]{preprocess_ww_data()}} contains the required columns. If
it does not, we want to tell the user which column is missing. This will not
however, ensure that the element sof the column are of the right type,
or check that the values of them make sense.
}
