% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_rt_length}
\alias{check_rt_length}
\title{Check that R(t) specified for generating simulated data is of sufficient
length}
\usage{
check_rt_length(
  r_in_weeks,
  ot,
  nt,
  forecast_horizon,
  call = rlang::caller_env()
)
}
\arguments{
\item{r_in_weeks}{a vector indicating the R(t) in weeks}

\item{ot}{integer indicating the observed time: length of hospital admissions
calibration time in days}

\item{nt}{integer indicating the nowcast time: length of time between last
hospital admissions date and forecast date in days}

\item{forecast_horizon}{integer indicating the duration of the forecast in
days e.g. 28 days}

\item{call}{Calling environment to be passed to the type checker}
}
\value{
NULL, invisible
}
\description{
Check that R(t) specified for generating simulated data is of sufficient
length
}
