% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_site_and_lab_indices}
\alias{check_site_and_lab_indices}
\title{Check that the specified site and lab indices line up}
\usage{
check_site_and_lab_indices(site, lab, call = rlang::caller_env())
}
\arguments{
\item{site}{vector of integers indicating which site (WWTP) each separate
lab-site observation comes frm}

\item{lab}{ector of integers indicating which lab the lab-site observations
come from}

\item{call}{}
}
\value{
NULL, invisibly
}
\description{
Check that the specified site and lab indices line up
}
