% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_vector}
\alias{check_vector}
\title{Check that the argument is a vector}
\usage{
check_vector(x, arg = "x", call = rlang::caller_env())
}
\arguments{
\item{x}{Object with type checking, here this will be a vector}

\item{arg}{Name of the argument supplying the object}

\item{call}{Calling environment to be passed to \code{\link[cli:cli_abort]{cli::cli_abort()}}}
}
\value{
NULL, invisibly
}
\description{
Check that the argument is a vector
}
