% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_ww_site_pops}
\alias{check_ww_site_pops}
\title{Check that the sum of the wastewater site populations don't exceed
the total population}
\usage{
check_ww_site_pops(pop_size, ww_pop_sites, call = rlang::caller_env())
}
\arguments{
\item{pop_size}{integer indicating the population size in the hypothetical
state}

\item{ww_pop_sites}{vector indicating the population size in the
catchment area in each of those sites}

\item{call}{Calling environment to be passed to the type checker}
}
\value{
NULL, invisibly
}
\description{
Check that the sum of the wastewater site populations don't exceed
the total population
}
