% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_logsd}
\alias{convert_to_logsd}
\title{Get the sd of a Normal distribution for a random variable Y
needed to ensure that the distribution of X = exp(Y) (which is Log-Normal)
has a specified mean and sd.}
\usage{
convert_to_logsd(mean, sd)
}
\arguments{
\item{mean}{target mean for the Log-Normal distribution of X}

\item{sd}{target sd for the Log-Normal distribution of X}
}
\value{
corresponding sd for the underlying Normal distribution of Y = log(X)
}
\description{
see arithmetic moments here
https://en.wikipedia.org/wiki/Log-normal_distribution
}
