% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_corr_matrix_func.R
\name{custom_corr_matrix_func}
\alias{custom_corr_matrix_func}
\title{Passes user specified correlation matrix, used for \code{generate_simulated_data}.}
\usage{
custom_corr_matrix_func(corr_function_params = list(corr_matrix = NULL))
}
\arguments{
\item{corr_function_params}{Structured as follows : {corr_matrix}.
Ensure that this list follows this structure. Correlation matrix with proprer
dimensions.  Defaults are set to NULL.}
}
\value{
The same correlation matrix that was passed in.
}
\description{
Passes user specified correlation matrix, used for \code{generate_simulated_data}.
}
