% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_diagnostics.R
\name{diagnostic_summary}
\alias{diagnostic_summary}
\title{Method for printing the CmdStan diagnostic summary
for a wwinference_fit object}
\usage{
diagnostic_summary(ww_fit, ...)
}
\arguments{
\item{ww_fit}{An object of class wwinference_fit}

\item{...}{additional arguments}
}
\description{
Method for printing the CmdStan diagnostic summary
for a wwinference_fit object
}
\seealso{
Other diagnostics: 
\code{\link{get_model_diagnostic_flags}()},
\code{\link{parameter_diagnostics}()},
\code{\link{wwinference}()}
}
\concept{diagnostics}
