% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_component_fwd_sim.R
\name{downsample_for_frequency}
\alias{downsample_for_frequency}
\title{Downsample the predicted wastewater concentrations based on the
lab site reporting frequency}
\usage{
downsample_for_frequency(
  log_conc_lab_site,
  n_lab_sites,
  ot,
  ht,
  lab_site_reporting_freq
)
}
\arguments{
\item{log_conc_lab_site}{The matrix of n_lab_sites by n time points
indicating the underlying expected  observed concentrations}

\item{n_lab_sites}{Integer indicating the number of unique lab-site
combinations}

\item{ot}{integer indicating the number of days we will have observed data
for in the calibration period}

\item{ht}{integer indicating the time after the last observed time to
the end of the forecast time}

\item{lab_site_reporting_freq}{vector indicating the mean frequency of
wastewater measurements in each site per day (e.g. 1/7 is once per week)}
}
\value{
A sparse matrix of \code{n_lab_sites} rows and \code{ot} + \code{ht} columns of
but with NAs for when observations are not measured/reported.
}
\description{
Downsample the predicted wastewater concentrations based on the
lab site reporting frequency
}
