% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_decay_corr_func.R
\name{exponential_decay_corr_func}
\alias{exponential_decay_corr_func}
\title{Generates the correlation matrix for an exponential
decay correlation function.}
\usage{
exponential_decay_corr_func(
  corr_function_params = list(dist_matrix = NULL, phi = NULL, l = NULL)
)
}
\arguments{
\item{corr_function_params}{Sturcuted as follows : {dist.matrix, phi, l}.}
}
\value{
Correlation matrix.
}
\description{
Generates the correlation matrix for an exponential
decay correlation function.
}
