% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exponential_decay_corr_func_r.R
\name{exponential_decay_corr_func_r}
\alias{exponential_decay_corr_func_r}
\title{Generates the correlation matrix for an exponential
decay correlation function.}
\usage{
exponential_decay_corr_func_r(
  corr_function_params = list(dist_matrix = NULL, phi = NULL, l = NULL)
)
}
\arguments{
\item{corr_function_params}{Structured as follows : dist_matrix, phi, l.
Ensure that this list follows this structure. Distance matrix is constructed
externally, phi is a numeric value, l is also a numeric value.  Defaults are
set to NULL.}
}
\value{
Correlation matrix.
}
\description{
Generates the correlation matrix for an exponential
decay correlation function.
}
