% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_component_fwd_sim.R
\name{format_hosp_data}
\alias{format_hosp_data}
\title{Format the hospital admissions data into a tidy dataframe}
\usage{
format_hosp_data(pred_obs_hosp, dur_obs, pop_size, date_df)
}
\arguments{
\item{pred_obs_hosp}{vector of non-negative integers indicating the number
of hospital admissions on each day}

\item{dur_obs}{integer indicating the number of days we want the
observations for}

\item{pop_size}{integer indicating the population size of the admissions
catchment area}

\item{date_df}{tibble of columns \code{date} and \code{t} that map time in days to
dates}
}
\value{
a tidy dataframe containing counts of admissions by date alongside
population size
}
\description{
Format the hospital admissions data into a tidy dataframe
}
