% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_component_fwd_sim.R
\name{format_subpop_hosp_data}
\alias{format_subpop_hosp_data}
\title{Format the subpopulation-level hospital admissions data into a tidy
dataframe}
\usage{
format_subpop_hosp_data(pred_obs_hosp_subpop, dur_obs, subpop_map, date_df)
}
\arguments{
\item{pred_obs_hosp_subpop}{matrix of non-negative integers indicating the
number of hospital admissions on each day in each subpopulation. Rows are
subpopulations, columns are time points}

\item{dur_obs}{integer indicating the number of days we want the
observations for}

\item{subpop_map}{tibble mapping the numbered subpopulations to the
wastewater sites, must contain columns "subpop_index" and "subpop_name"}

\item{date_df}{tibble of columns \code{date} and \code{t} that map time in days to
dates}
}
\value{
a tidy dataframe containing counts of admissions by date alongside
population size for each subpopulation
}
\description{
Format the subpopulation-level hospital admissions data into a tidy
dataframe
}
