% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_component_fwd_sim.R
\name{format_ww_data}
\alias{format_ww_data}
\title{Format the wastewater data as a tidy data frame}
\usage{
format_ww_data(log_obs_conc_lab_site, ot, ht, date_df, site_lab_map)
}
\arguments{
\item{log_obs_conc_lab_site}{matrix of numeric values where rows are the
site-lab combinations and columns are the observed time points}

\item{ot}{integer indicating the number of days we will have observed data
for in the calibration period}

\item{ht}{integer indicating the time after the last observed epi indicator
and before the forecast date, of which there can still be wastewater
observations}

\item{date_df}{tibble of columns \code{date} and \code{t} that map time in days to
dates}

\item{site_lab_map}{tibble mapping sites, labs, lab site indices, and the
population size of the site}
}
\value{
a tidy dataframe containing the observed wastewater concentrations
in each site and lab at each time point
}
\description{
Format the wastewater data as a tidy data frame
}
