% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{generation_interval}
\alias{generation_interval}
\title{COVID-19 post-Omicron generation interval probability mass function}
\format{
An object of class \code{numeric} of length 15.
}
\source{
covid_pmfs.R
}
\usage{
generation_interval
}
\description{
\describe{
A vector that sums to 1, with each element representing the daily
probability of secondary onward transmission occurring on that day. The
first element of this vector represents the day after primary transmission
occurred, it is assumed to be impossible for primary and secondary
transmission to occur on the same day.
}
}
\keyword{datasets}
