% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stan_data.R
\name{get_count_data_sizes}
\alias{get_count_data_sizes}
\title{Get count data integer sizes for stan}
\usage{
get_count_data_sizes(
  input_count_data,
  forecast_date,
  forecast_horizon,
  calibration_time,
  last_count_data_date,
  uot,
  count_col_name = "count"
)
}
\arguments{
\item{forecast_date}{string indicating the forecast date}

\item{forecast_horizon}{integer indicating the number of days to make a
forecast for}

\item{calibration_time}{integer indicating the max duration in days that
the model is calibrated to hospital admissions for}

\item{last_count_data_date}{string indicating the date of the last observed
count data point}

\item{uot}{integer indicating the time of model initialization when there are
no observations}

\item{count_col_name}{A string represeting the name of the column in the
input_count_data that indicates the number of daily counts,
default is \code{count}}

\item{input_cout_data}{a dataframe with the input count data}
}
\value{
A list containing the integer sizes of the follow variables that
the stan model requires:
ht:  integer indicating horizon time for the model(hospital admissions
nowcast + forecast time in days)
ot: integer indicating the total duration of time that model for producing
counts (e.g. cases or admissions) has available calibration data
oht: integer indicating the number of count observations
n_weeks: number of weeks (rounded up) that counts are generated
from the model
tot_weeks: number of week(rounded up) that infections are generated for
}
\description{
Get count data integer sizes for stan
}
