% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stan_data.R
\name{get_count_indices}
\alias{get_count_indices}
\title{Get count data indices}
\usage{
get_count_indices(input_count_data)
}
\arguments{
\item{input_count_data}{a dataframe with the input count data}
}
\value{
A list containing the vectors of indices that
the stan model requires:
count_times: a vector of integer times corresponding to the times when the
count observations were made
}
\description{
Get count data indices
}
