% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stan_data.R
\name{get_date_time_spine}
\alias{get_date_time_spine}
\title{Produce a tibble with a date column and a 1-indexed t column}
\usage{
get_date_time_spine(first_date, last_date)
}
\arguments{
\item{first_date}{first date in the table. Must be coercible to a
date via \code{\link[=as.Date]{as.Date()}}.}

\item{last_date}{ldate date in the table. Must be coercible to a
date via \code{\link[=as.Date]{as.Date()}}.}
}
\value{
A tibble mapping dates to (1-indexed) model time.
}
\description{
For a range of dates, inclusive.
}
