% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stan_data.R
\name{get_first_calibration_date}
\alias{get_first_calibration_date}
\title{Get the first date of calibration count data}
\usage{
get_first_calibration_date(count_data, calibration_time)
}
\arguments{
\item{count_data}{dataframe with the input count data. Must contain a
column named \code{date}.}

\item{calibration_time}{Maximum number of days of count date to
use when fitting the model.}
}
\value{
The first calibration date.
}
\description{
Get the first date of calibration count data
}
\examples{

get_first_calibration_date(wwinference::hosp_data, 50)

get_first_calibration_date(wwinference::hosp_data, 70)
}
