% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ind_m}
\alias{get_ind_m}
\title{Get index matrix}
\usage{
get_ind_m(n_days, n_weeks)
}
\arguments{
\item{n_days}{number of days we will expand to}

\item{n_weeks}{number of weeks those days correspond to}
}
\value{
a n_day x n_week matrix for multiplying by weekly estimated
value to broadcast it to daily
}
\description{
Get a matrix to broadcast a vector from weekly to daily
}
\examples{
ind_m <- get_ind_m(14, 2)
}
