% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{get_inits}
\alias{get_inits}
\title{Given a set of prior parameters and stan data, initialize the model
near the center of the prior distribution}
\usage{
get_inits(stan_data, params)
}
\arguments{
\item{stan_data}{a list of data elements that will be passed to stan}

\item{params}{a dataframe of parameter values that are passed to stan
to specify the priors in the model}
}
\value{
a list of initial values for each of the parameters in the
\code{wwinference} model
}
\description{
Given a set of prior parameters and stan data, initialize the model
near the center of the prior distribution
}
