% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{get_inits_for_one_chain}
\alias{get_inits_for_one_chain}
\title{Given a set of prior parameters and stan data, initialize the model
near the center of the prior distribution}
\usage{
get_inits_for_one_chain(stan_data, stdev = 0.01)
}
\arguments{
\item{stan_data}{a list of data elements that will be passed to stan}

\item{stdev}{a numeric value indicating the standard deviation to sample
from when initializing, particularly from a standard normal. Also acts as
a multiplier on the prior standard deviation, to restrict the initial value
to be sampled from the center of the prior. Default is 0.01}
}
\value{
a list of initial values for each of the parameters in the
\code{wwinference} model
}
\description{
Given a set of prior parameters and stan data, initialize the model
near the center of the prior distribution
}
