% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stan_data.R
\name{get_input_count_data_for_stan}
\alias{get_input_count_data_for_stan}
\title{Get the input count data to pass directly to stan}
\usage{
get_input_count_data_for_stan(preprocessed_count_data, date_time_spine)
}
\arguments{
\item{preprocessed_count_data}{a dataframe with the input count data, must
have a column named \code{date}.}

\item{date_time_spine}{table mapping dates to model time index, as the
output of \code{\link[=get_date_time_spine]{get_date_time_spine()}}.}
}
\value{
data frame of count data passed to stan
}
\description{
Get the input count data to pass directly to stan
}
