% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stan_data.R
\name{get_input_ww_data_for_stan}
\alias{get_input_ww_data_for_stan}
\title{Get the input ww data passed directly to stan}
\usage{
get_input_ww_data_for_stan(
  preprocessed_ww_data,
  first_count_data_date,
  last_count_data_date,
  calibration_time,
  forecast_date
)
}
\arguments{
\item{preprocessed_ww_data}{a dataframe with the input wastewater data with
no gaps, must have the following columns: date, site, lab,
genome_copies_per_ml, site_pop, below_lod, and if removing outliers,
flag_as_ww_outlier}

\item{first_count_data_date}{The earliest day with an observation in the '
count dataset, in ISO8601 format (YYYY-MM-DD)}

\item{last_count_data_date}{the last date that a count observation is
presen, in ISO8601 format (YYYY-MM-DD)}

\item{calibration_time}{integer indicating the max duration in days that
the model is calibrated to the count data for}

\item{forecast_date}{a character string in ISO8601 format (YYYY-MM-DD)
indicating the date that the forecast is to be made. Default is NULL}
}
\value{
dataframe of the ww data passed to stan
}
\description{
Get the input ww data passed directly to stan
}
