% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stan_data.R
\name{get_lab_site_site_spine}
\alias{get_lab_site_site_spine}
\title{Get mapping from lab-site to site}
\usage{
get_lab_site_site_spine(input_ww_data)
}
\arguments{
\item{input_ww_data}{a dataframe of the wastewater data to be passed
directly to stan, must have the following columns: date, site, lab,
genome_copies_per_ml, site_pop, below_lod, and exclude}
}
\value{
a dataframe mapping the unique combinations of sites and labs
to their indices in the model and the population of the site in that
observation unit (lab_site)
}
\description{
Get mapping from lab-site to site
}
