% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stan_data.R
\name{get_lab_site_subpop_spine}
\alias{get_lab_site_subpop_spine}
\title{Get lab-site subpopulation spine}
\usage{
get_lab_site_subpop_spine(lab_site_site_spine, site_subpop_spine)
}
\arguments{
\item{lab_site_site_spine}{tibble mapping lab-sites to sites}

\item{site_subpop_spine}{tibble mapping sites to subpopulations}
}
\value{
a tibble mapping lab-sites to subpopulations
}
\description{
Get lab-site subpopulation spine
}
