% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stan_data.R
\name{get_last_target_date}
\alias{get_last_target_date}
\title{Get the final target date of the forecast period}
\usage{
get_last_target_date(forecast_date, forecast_horizon)
}
\arguments{
\item{forecast_date}{Forecast date, coercible by \code{\link[=as.Date]{as.Date()}}.}

\item{forecast_horizon}{Forecast horizon, in days.}
}
\value{
The last target date, i.e. the last date
for which the model will produce a forecast.
}
\description{
Enforces the convention that the forecast_date itself
is day 0 of the forecast horizon, not (e.g. day 1).
}
\examples{
get_last_target_date("2026-01-01", 3)
get_last_target_date("2026-01-01", 0)
}
