% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{get_plot_subpop_rt}
\alias{get_plot_subpop_rt}
\title{Get plot of fit, nowcasted, and forecasted R(t) in each subpopulation}
\usage{
get_plot_subpop_rt(draws, forecast_date)
}
\arguments{
\item{draws}{A dataframe containing the posterior draws with the data joined
to it. This is \code{draws_df} output of the call to \code{wwinference()}}

\item{forecast_date}{A string indicating the date we made the forecast, for
plotting, in ISO8601 format YYYY-MM-DD}
}
\value{
A ggplot object containing faceted plots of the R(t) estimate in each
subpopulation (so wastewater sites + those not on wastewater)
}
\description{
Get plot of fit, nowcasted, and forecasted R(t) in each subpopulation
}
