% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{get_plot_ww_conc}
\alias{get_plot_ww_conc}
\title{Get plot of fit and forecasted wastewater concentrations}
\usage{
get_plot_ww_conc(draws, forecast_date, n_draws_to_plot = 100)
}
\arguments{
\item{draws}{A dataframe containing the posterior draws with the data joined
to it. This is the \code{draws_df} output of a call to \code{\link[=wwinference]{wwinference()}}. It
expects the following column names: \code{date}, \code{pred_value}, \code{draw},
and \code{name}.}

\item{forecast_date}{A string indicating the date we made the forecast, for
plotting, in ISO8601 format YYYY-MM-DD}

\item{n_draws_to_plot}{An integer indicating how many draws from the
posterior to include in the plot, default is \code{100}}
}
\value{
a ggplot object containing faceted plots of the wastewaster
concentrations in each site and lab combination
}
\description{
Get plot of fit and forecasted wastewater concentrations
}
