% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_component_fwd_sim.R
\name{get_pred_obs_conc}
\alias{get_pred_obs_conc}
\title{Get the predicted concentrations in each lab site}
\usage{
get_pred_obs_conc(
  n_lab_sites,
  ot,
  ht,
  log_g_over_n_site,
  log_m_lab_sites,
  sigma_ww_lab_site,
  site,
  ml_of_ww_per_person_day
)
}
\arguments{
\item{n_lab_sites}{Integer indicating the number of unique combinations of
labs and sites in the dataset.}

\item{ot}{integer indicating the number of days we will have observed data
for in the calibration period}

\item{ht}{integer indicating the time after the last observed time to}

\item{log_g_over_n_site}{matrix of n_site rows and ot + ht columns indicating
the genomes per person each day in each site}

\item{log_m_lab_sites}{vector of the lab-site mutlipliers}

\item{site}{vector of integers indicating which site (WWTP) each separate
lab-site observation comes from}

\item{ml_of_ww_per_person_day}{Scalar indicating the number of mL of
wastewater produced per person per day}
}
\value{
A matrix of \code{n_lab_sites} rows and \code{ot} + \code{ht} columns indcating the
predicted concentration of wastewater in each lab-site and day
}
\description{
Get the predicted concentrations in each lab site
}
