% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_component_fwd_sim.R
\name{get_pred_subpop_gen_per_n}
\alias{get_pred_subpop_gen_per_n}
\title{Get the predicted genomes per person in each subpopulation}
\usage{
get_pred_subpop_gen_per_n(
  convolve_fxn,
  n_sites,
  uot,
  ot,
  ht,
  new_i_over_n_site,
  log10_g_prior_mean,
  vl_trajectory
)
}
\arguments{
\item{convolve_fxn}{function used to convolve infections with delay pmf.
This will typically take \code{convolve_dot_product()} from the
\link{wwinference.stan} model}

\item{n_sites}{integer indicating the number of sites}

\item{uot}{integer indicating the days for exponential growth initialization
to occur (referred to as unobserved time)}

\item{ot}{integer indicating the number of days we will have observed data
for in the calibration period}

\item{ht}{integer indicating the time after the last observed time to}

\item{new_i_over_n_site}{matrix of n_subpops rows x n total time points
columns indicating the per capita infections in each subpopulation}

\item{log10_g_prior_mean}{log10 scale prior on the number of genomes shed
per infection}

\item{vl_trajectory}{vector of simplex indicating the proportion of all
shedding occuring on each day after infection}
}
\value{
A matrix of \code{n_sites} rows and \code{ot}+ \code{ht} columns indicating
the predicted number of genomes per person per day in each site each day
}
\description{
Get the predicted genomes per person in each subpopulation
}
