% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stan_data.R
\name{get_site_subpop_spine}
\alias{get_site_subpop_spine}
\title{Get site to subpopulation map}
\usage{
get_site_subpop_spine(input_ww_data, input_count_data)
}
\arguments{
\item{input_ww_data}{a dataframe of the wastewater data to be passed
directly to stan, must have the following columns: date, site, lab,
genome_copies_per_ml, site_pop, below_lod, and exclude}

\item{input_count_data}{a dataframe of the count data to be passed
directly to stan, , must have the following columns: date, count, total_pop}
}
\value{
a dataframe mapping the sites to the corresponding subpopulation and
subpopulation index, plus the population in each subpopulation. Imposes
the logic to add a subpopulation if the total population is greater than
the sum of the site populations in the input wastewater data
}
\description{
Get site to subpopulation map
}
