% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_component_fwd_sim.R
\name{get_time_varying_daily_ihr}
\alias{get_time_varying_daily_ihr}
\title{Get time varying IHR}
\usage{
get_time_varying_daily_ihr(p_hosp_mean, uot, ot, ht, tot_weeks, p_hosp_w_sd_sd)
}
\arguments{
\item{p_hosp_mean}{mean probability of hospital admission given infection}

\item{uot}{integer indicating the days for exponential growth initialization
to occur (referred to as unobserved time)}

\item{ot}{integer indicating the number of days we will have observed data
for in the calibration period}

\item{ht}{integer indicating the time after the last observed time to
forecast}

\item{tot_weeks}{integer indicating the total number of weeks including the
\code{uot}}

\item{p_hosp_w_sd_sd}{the standard deviation in logit scale of the
IHR}
}
\value{
a vector of daily IHR values
}
\description{
Get time varying IHR
}
