% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stan_data.R
\name{get_ww_indices_and_values}
\alias{get_ww_indices_and_values}
\title{Get wastewater indices and values for stan}
\usage{
get_ww_indices_and_values(
  input_ww_data,
  date_time_spine,
  lab_site_site_spine,
  site_subpop_spine,
  lab_site_subpop_spine
)
}
\arguments{
\item{input_ww_data}{tibble with the input wastewater data and indices
needed for stan}

\item{date_time_spine}{tibble mapping dates to time in days}

\item{lab_site_site_spine}{tibble mapping lab-sites to sites}

\item{site_subpop_spine}{tibble mapping sites to subpopulations}

\item{lab_site_subpop_spine}{tibble mapping lab-sites to subpopulations}
}
\value{
a list of the vectors needed for stan
}
\description{
Get wastewater indices and values for stan
}
