% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stan_data.R
\name{get_ww_values}
\alias{get_ww_values}
\title{Get wastewater data values}
\usage{
get_ww_values(
  ww_data,
  ww_measurement_col_name = "genome_copies_per_ml",
  ww_lod_value_col_name = "lod",
  ww_site_pop_col_name = "site_pop",
  one_pop_per_site = TRUE
)
}
\arguments{
\item{ww_data}{Input wastewater dataframe containing one row
per observation, with outliers already removed}

\item{ww_measurement_col_name}{A string representing the name of the column
in the input_ww_data that indicates the wastewater measurement value in
natural scale, default is \code{genome_copies_per_ml}}

\item{ww_lod_value_col_name}{A string representing the name of the column
in the ww_data that indicates the value of the LOD in natural scale,
default is \code{lod}}

\item{ww_site_pop_col_name}{A string representing the name of the column in
the ww_data that indicates the number of people represented by that
wastewater catchment, default is \code{site_pop}}

\item{one_pop_per_site}{a boolean variable indicating if there should only
be on catchment area population per site, default is \code{TRUE} because this is
what the stan model expects}
}
\value{
A list containing the necessary vectors of values that
the stan model requires:
ww_lod: a vector of the LODs of the corresponding wastewater measurement
pop_ww: a vector of the population sizes of the wastewater catchment areas
in order of the sites by site_index
log_conc: a vector of the log of the wastewater concentration observation
}
\description{
Get wastewater data values
}
