% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hosp_data_eval}
\alias{hosp_data_eval}
\title{Example hospital admissions dataset for evaluation}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 127 rows and 3 columns.
}
\source{
vignette_data.R
}
\usage{
hosp_data_eval
}
\description{
A dataset containing the simulated daily hospital admissions that the model
will be evaluated against (labeled here as \code{daily_hosp_admits_for_eval})
by date of admission (\code{date}). This data is not needed to fit the model,
but is used in the Getting Started vignette to demonstrate the forecasted
hospital admissions compared to those later observed.
}
\details{
This data is generated via the default values in the
\code{generate_simulated_data()} function.

The variables are as follows:
\describe{
\item{date}{Date the hospital admissions occurred, formatte din ISO8601
standatds as YYYY-MM-DD}
\item{daily_hosp_admits_for_eval}{The number of individuals admitted to the
hospital on that date, available beyond the forecast date for evaluating
the forecasted hospital admissions}
\item{state_pop}{The number of people contributing to the daily hospital
admissions}
}
}
\keyword{datasets}
