% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hosp_data_eval_ind}
\alias{hosp_data_eval_ind}
\title{Example subpopulation level hospital admissions dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 127 rows and 3 columns.
}
\source{
vignette_data.R
}
\usage{
hosp_data_eval_ind
}
\description{
A dataset containing the simulated daily hospital admissions
(labeled here as \code{daily_hosp_admits}) by date of admission (\code{date}) in
each subpopulation.
Additional columns that are the population size of the
population contributing to the hospital admissions. In this instance,
the subpopulations here are each of the wastewater catchment areas plus
an additional subpopulation for the portion of the population not captured
by wastewater surveillance. The data generated are daily hospital
admissions but they could be any other epidemiological count dataset e.g.
cases. This data should only contain hospital admissions that would have
been available as of the date that the forecast was made.
}
\details{
This data is generated via the default values in the
\code{generate_simulated_data()} function.

The variables are as follows:
\describe{
\item{date}{Date the hospital admissions occurred, formatte din ISO8601
standatds as YYYY-MM-DD}
\item{daily_hosp_admits_for_eval}{The number of individuals admitted to the
hospital on that date, available beyond the forecast date for evaluating
the forecasted hospital admissions}
\item{state_pop}{The number of people contributing to the daily hospital
admissions}
}
}
\keyword{datasets}
