% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independence_corr_func.R
\name{independence_corr_func}
\alias{independence_corr_func}
\title{Generates the correlation matrix for a non-spatial process.}
\usage{
independence_corr_func(corr_function_params = list(num_sites = NULL))
}
\arguments{
\item{corr_function_params}{NULL}
}
\value{
Correlation matrix of diagonal ones.
}
\description{
Generates the correlation matrix for a non-spatial process.
}
