% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{indicate_ww_exclusions}
\alias{indicate_ww_exclusions}
\title{Indicate data that we want to exclude from model fitting}
\usage{
indicate_ww_exclusions(
  data,
  outlier_col_name = "flag_as_ww_outlier",
  remove_outliers = TRUE
)
}
\arguments{
\item{data}{A dataframe of preprocessed data to be used to fit the o}

\item{outlier_col_name}{A character string indicating the name of the column
containing the outlier indicator, must contain only 0 or 1}

\item{remove_outliers}{A boolean indicating whether or not to exclude the
outliers from the fitting. If TRUE, copy outliers to exclusions, if FALSE,
set exclusions to none}
}
\value{
a dataframe with the same columns as in \code{data} plus an additional
\code{exclude} column containing 0s for the data to be passed to the model
and 1s where the data should be excluded
}
\description{
This function takes in a dataframe which contains an outlier
column name specified by the \code{outlier_col_name}.
}
\examples{
data <- tibble::tibble(
  date = c("2023-10-01", "2023-10-02"),
  genome_copies_per_mL = c(300, 3e6),
  flag_as_ww_outlier = c(0, 1),
  exclude = c(0, 0)
)
data_w_exclusions <- indicate_ww_exclusions(data,
  outlier_col_name = "flag_as_ww_outlier",
  remove_outliers = TRUE
)
}
