% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inf_to_hosp}
\alias{inf_to_hosp}
\title{COVID-19 time delay distribution from infection to hospital admission}
\format{
An object of class \code{numeric} of length 55.
}
\source{
covid_pmfs.R
}
\usage{
inf_to_hosp
}
\description{
\describe{
A vector that sums to 1, with each element representing the daily
probabilty of transitioning from infected to hospitalized, conditioned on
being infected and eventually ending up hospitalized. The first element
represents the probability of being infected and admitted to the hospital
on the same day
}
}
\keyword{datasets}
