% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_distribs.R
\name{make_incubation_period_pmf}
\alias{make_incubation_period_pmf}
\title{Make incubation period pmf}
\usage{
make_incubation_period_pmf(
  backward_scale = 3.6,
  backward_shape = 1.5,
  r = 0.15
)
}
\arguments{
\item{backward_scale}{numeric indicating the scale parameter for the Weibull
used in producing the incubation period distribution. default is \code{3.60} for
COVID}

\item{backward_shape}{numeric indicating the shape parameter for the Weibull
used in producing the incubation period distribution, default is \code{1.50} for
COVID}

\item{r}{numeric indicating the exponential rate used in producing the
correction on the incubaion period distribution, default is \code{0.15} for COVID}
}
\value{
pmf of incubation period
}
\description{
When the default arguments are used, this returns a pmf
corresponding to the incubation period for COVID after Omicron used in
Park et al 2023. These estimates are from early Omicron.
}
