% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_distribs.R
\name{make_reporting_delay_pmf}
\alias{make_reporting_delay_pmf}
\title{Make reporting delay pmf}
\usage{
make_reporting_delay_pmf(incubation_period_pmf, hospital_onset_delay_pmf)
}
\arguments{
\item{incubation_period_pmf}{a numeric vector, sums to 1, indicating
the probability of time from infection to symptom onset}

\item{hospital_onset_delay_pmf}{a numeric vector, sums to 1, indicating the
proabbility of time from symptom onset to hospital admissions}
}
\value{
convolution of incubation period and sympton onset to hospital
admission pmf
}
\description{
Convolve the incubation period pmf with the symptom to hospital admission pmf
and normalize
}
\examples{
inc_pmf <- make_incubation_period_pmf(3.6, 1.5, 0.15)
hosp_delay_pmf <- make_hospital_onset_delay_pmf(7, 2.5)
inf_to_hosp_pmf <- make_reporting_delay_pmf(inc_pmf, hosp_delay_pmf)
}
