% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_weekly_data.R
\name{make_weekly_data}
\alias{make_weekly_data}
\title{Make daily data into weekly data
This is an internal function used to generate simulated weekly data}
\usage{
make_weekly_data(hosp_data, day_of_week_to_sum = "Saturday")
}
\arguments{
\item{hosp_data}{A tibble containing daily hospital admissions data,
expects the following columns: "date", "daily_hosp_admits"}

\item{day_of_week_to_sum}{A character string indicating what day of the
week to assign the past 7 days of hospital admissions to. Must be full
weekday name with first letter in uppercase}
}
\description{
Make daily data into weekly data
This is an internal function used to generate simulated weekly data
}
