% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_diagnostics.R
\name{parameter_diagnostics}
\alias{parameter_diagnostics}
\title{Method for printing the CmdStan parameter diagnostics for a
wwinference_fit_object}
\usage{
parameter_diagnostics(ww_fit, ...)
}
\arguments{
\item{ww_fit}{An object of class wwinference_fit}

\item{...}{additional arguments}
}
\description{
Method for printing the CmdStan parameter diagnostics for a
wwinference_fit_object
}
\seealso{
Other diagnostics:
\code{\link[=get_model_diagnostic_flags]{get_model_diagnostic_flags()}},
\code{\link[=summary_diagnostics]{summary_diagnostics()}},
\code{\link[=wwinference]{wwinference()}}
}
\concept{diagnostics}
