% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocess_hosp_data}
\alias{preprocess_hosp_data}
\title{Get input hospital admissions data}
\usage{
preprocess_hosp_data(
  hosp_data,
  forecast_date,
  count_col_name = "daily_hosp_admits",
  pop_size_col_name = "state_pop"
)
}
\arguments{
\item{hosp_data}{dataframe containing the following columns: date,
a count column, and a population size column}

\item{forecast_date}{The forecast date for this iteration,
formatted as a character string in IS08601 format (YYYY-MM-DD).}

\item{count_col_name}{name of the column containing the epidemiological
indicator, default is \code{daily_hosp_admits}}

\item{pop_size_col_name}{name of the column containing the population size
of that the counts are coming from, default is \code{state_pop}}
}
\value{
a dataframe containing the hospital admissions data renamed to
have the following columns \code{date}, \code{count}, \code{total_pop} and \code{forecast_date}
}
\description{
Get input hospital admissions data
}
\examples{
hosp_data_preprocessed <- preprocess_hospdata(hosp_data, "2023-12-01")
}
