% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocess_ww_data}
\alias{preprocess_ww_data}
\title{Get input wastewater data}
\usage{
preprocess_ww_data(
  ww_data,
  forecast_date,
  conc_col_name = "genome_copies_per_ml"
)
}
\arguments{
\item{ww_data}{dataframe containing the following columns: site, lab,
date, a column for concentration, and lod}

\item{forecast_date}{The forecast date for this iteration,
formatted as a character string in IS08601 format (YYYY-MM-DD).}

\item{conc_col_name}{name of the column containing the concentration
measurements in the wastewater data, default is \code{genome_copies_per_ml}}
}
\value{
a dataframe containing the transformed and clean NWSS data
at the site and lab label for the forecast date and location specified
}
\description{
Get input wastewater data
}
\examples{
ww_data_preprocessed <- preprocess_ww_data(ww_data, "2023-12-01")
}
