% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wwinference.R
\name{print.wwinference_fit}
\alias{print.wwinference_fit}
\title{Printing method for object of class \code{wwinference_fit}}
\usage{
\method{print}{wwinference_fit}(x, ...)
}
\arguments{
\item{x}{Object of class \code{wwinference.fit}}

\item{...}{Additional parameters}
}
\description{
Prints the labeled elements of the \code{wwinference_fit} object
}
