% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_corr_matrix_func.R
\name{rand_corr_matrix_func}
\alias{rand_corr_matrix_func}
\title{Generates a random correlation matrix.}
\usage{
rand_corr_matrix_func(corr_function_params = list(n_sites = NULL, seed = NULL))
}
\arguments{
\item{corr_function_params}{Structured as follows : {n_sites, seed}.
Ensure that this list follows this structure. Number of sites n_sites, and
seed desired for rng.  Defaults are set to NULL.}
}
\value{
Correlation matrix.  A symmetric matrix with off diagonal values
sampled randomly from -1 to 1, using a uniform distribution.
}
\description{
Generates a random correlation matrix.
}
