% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rt_global_data}
\alias{rt_global_data}
\title{Example of Global reproduction number from corr. func. model.}
\format{
An object of class \code{numeric} of length 127.
}
\source{
vignette_data.R
}
\usage{
rt_global_data
}
\description{
\describe{A vector containing the global reproduction number for the
\code{hosp_data}, \code{ww_data}, and \code{hosp_data_eval} datasets.
This data is generated via the default values in the
\code{generate_simulation_data()} function.
}
}
\keyword{datasets}
