% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rt_site_data_ind}
\alias{rt_site_data_ind}
\title{Example of Site reproduction number from spatially independent model.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 5 rows and 127 columns.
}
\source{
vignette_data.R
}
\usage{
rt_site_data_ind
}
\description{
\describe{A matrix containing the global reproduction number for the
\code{hosp_data_ind}, \code{ww_data_ind}, and \code{hosp_data_eval_ind} datasets.
Rows are sites, and columns are time.
This data is generated via the default values in the
\code{generate_simulation_data()} function.
}
}
\keyword{datasets}
