% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_rt_process.R
\name{spatial_rt_process}
\alias{spatial_rt_process}
\title{Generates the subpopulation Rt trajectories for a spatially correlated model}
\usage{
spatial_rt_process(
  log_state_rt,
  corr_function,
  corr_function_params,
  phi_rt,
  sigma_eps
)
}
\arguments{
\item{corr_function}{Function for the correlation structure desired.}

\item{corr_function_params}{List of parameters of desired correlation
function, should contain dist matrix.}

\item{phi_rt}{Coefficient for AR(1) temporal correlation on
subpopulation deviations.}

\item{sigma_eps}{Parameter for construction of covariance matrix
of spatial epsilon.}

\item{state_rt}{"State" level unadjusted Rt, on log scale.}
}
\value{
A matrix for unadjusted Rt where rows are subpopulations
and columns are time points.
}
\description{
Generates the subpopulation Rt trajectories for a spatially correlated model
}
