% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subpop_hosp_data}
\alias{subpop_hosp_data}
\title{Example subpopulation level hospital admissions dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 450 rows and 4 columns.
}
\source{
vignette_data.R
}
\usage{
subpop_hosp_data
}
\description{
A dataset containing the simulated daily hospital admissions
(labeled here as \code{daily_hosp_admits}) by date of admission (\code{date}) in
each subpopulation.
Additional columns that are the population size of the
population contributing to the hospital admissions. In this instance,
the subpopulations here are each of the wastewater catchment areas plus
an additional subpopulation for the portion of the population not captured
by wastewater surveillance. The data generated are daily hospital
admissions but they could be any other epidemiological count dataset e.g.
cases. This data should only contain hospital admissions that would have
been available as of the date that the forecast was made.
}
\details{
This data is generated via the default values in the
\code{generate_simulated_data()} function.

The variables are as follows:
\describe{
\item{date}{Date the hospital admissions occurred, formatted in ISO8601
standards as YYYY-MM-DD}
\item{subpop_name}{A string indicating the subpopulation the hospital
admissiosn corresponds to. This is either a wastewater site, or the
remainder of the population}
\item{daily_hosp_admits}{The number of individuals admitted to the
hospital on that date, available as of the forecast date}
\item{subpop_pop}{The number of people contributing to the daily hospital
admissions in each subpopulation}
}
}
\keyword{datasets}
