% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subpop_hosp_data_eval}
\alias{subpop_hosp_data_eval}
\title{Example subpopulation level retrospective hospital admissions dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 635 rows and 4 columns.
}
\source{
vignette_data.R
}
\usage{
subpop_hosp_data_eval
}
\description{
A dataset containing the simulated daily hospital admissions
(labeled here as \code{daily_hosp_admits}) by date of admission (\code{date}) in
each subpopulation observed retrospectively.
Additional columns that are required are the population size of the
population contributing to the hospital admissions. In this instance,
the subpopulations here are each of the wastewater catchment areas plus
an additional subpopulation for the portion of the population not captured
by wastewater surveillance. The data generated are daily hospital
admissions but they could be any other epidemiological count dataset e.g.
cases.This data should contain hospital admissions retrospectively beyond
the forecast date in order to evaluate the forecasts.
}
\details{
This data is generated via the default values in the
\code{generate_simulated_data()} function. They represent the bare minimumum
required fields needed to pass to the model, and we recommend that users
try to format their own data to match this format.

The variables are as follows:
\describe{
\item{date}{Date the hospital admissions occurred, formatted in ISO8601
standards as YYYY-MM-DD}
\item{subpop_name}{A string indicating the subpopulation the hospital
admissions corresponds to. This is either a wastewater site, or the
remainder of the population}
\item{daily_hosp_admits_for_eval}{The number of individuals admitted to the
hospital on that date, available as of the forecast date}
\item{subpop_pop}{The number of people contributing to the daily hospital
admissions in each subpopulation}
}
}
\keyword{datasets}
