% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_component_fwd_sim.R
\name{subpop_inf_process}
\alias{subpop_inf_process}
\title{Get the subpopulation level incident infections}
\usage{
subpop_inf_process(
  n_subpops,
  uot,
  ot,
  ht,
  unadj_r_site,
  initial_growth,
  initial_growth_prior_sd,
  i0_over_n,
  sd_i0_over_n,
  generation_interval,
  infection_feedback,
  infection_feedback_pmf
)
}
\arguments{
\item{n_subpops}{integer indicating the number of subpopulations}

\item{uot}{integer indicating the days for exponential growth initialization
to occur (referred to as unobserved time)}

\item{ot}{integer indicating the number of days we will have observed data
for in the calibration period}

\item{ht}{integer indicating the time after the last observed time to
forecast}

\item{unadj_r_site}{n_subpop x n_weeks matrix of the unadjusted site level
R(t)}

\item{initial_growth}{float indicating the mean of the initial growth rate
in the unobserved time}

\item{initial_growth_prior_sd}{float indicating the standard deviation on
the initial growth rate across subpopulations}

\item{i0_over_n}{float indicating the mean of the initial per capita
infections}

\item{sd_i0_over_n}{float indicating the standard deviation of the log of the
initial infections per capita across subpopulations}

\item{generation_interval}{vector of simplex describing the probability of
each time from infection to onwards transmission}

\item{infection_feedback}{boolean indicating whether or not to include
infection feedback}

\item{infection_feedback_pmf}{vector of simplex describing the delay from
incident infections to feedback on incident infections}
}
\value{
A list containing 3 outputs: i_n: n_subpop X total days matrix of
daily incident infections per capita in each subpopulation, r_site: adjusted
subpopulation level R(t) estimate, i_n_global: vector of daily incident
infections per capita in the global population
}
\description{
Get the subpopulation level incident infections
}
