% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_component_fwd_sim.R
\name{subpop_rt_process}
\alias{subpop_rt_process}
\title{Get subpopulation level R(t) estimates assuming time and space independence}
\usage{
subpop_rt_process(n_subpops, r_weeks, subpop_level_rt_variation)
}
\arguments{
\item{n_subpops}{integer indicating the number of subpopulations,
usually this will be n_sites + 1}

\item{r_weeks}{The "global" R(t) in weeks}

\item{subpop_level_rt_variation}{The standard deviation of the Gaussian
to generate deviation at the site level in the R(t) estimate}
}
\value{
A n_sites+1 by n_weeks matrix containing the subpopulation R(t)s
}
\description{
Get subpopulation level R(t) estimates assuming time and space independence
}
\seealso{
Other Model components:
\code{\link{calc_rt}()},
\code{\link{create_site_lab_map}()},
\code{\link{downsample_ww_obs}()},
\code{\link{format_hosp_data}()},
\code{\link{format_ww_data}()},
\code{\link{get_global_rt}()},
\code{\link{get_pred_obs_conc}()},
\code{\link{get_pred_subpop_gen_per_n}()},
\code{\link{get_time_varying_daily_ihr}()},
\code{\link{subpop_inf_process}()}
}
\concept{Model components}
