% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{throw_type_error}
\alias{throw_type_error}
\title{Throw an informative type error on a user-provided input}
\usage{
throw_type_error(object, arg_name, expected_type, call = rlang::caller_env())
}
\arguments{
\item{object}{Object with incorrect type}

\item{arg_name}{Name of the argument corresponding to \code{object}}

\item{expected_type}{The type that the user should provide instead}

\item{call}{The calling environment to be reflected in the error message}
}
\value{
This function is called for its side effect of throwing an error. It
should never return.
}
\description{
Follows the guidance from \code{\link[rlang:abort]{rlang::abort()}} on applying a call and a class
in the error message. Used as a base in type-checkers to throw a properly
formatted error.
}
