% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_simplex}
\alias{to_simplex}
\title{Normalize vector to a simplex}
\usage{
to_simplex(vector)
}
\arguments{
\item{vector}{numeric vector}
}
\value{
vector whos sum adds to 1
}
\description{
Normalize vector to a simplex
}
\examples{
to_simplex(c(1, 1, 1))
}
