% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{true_global_rt}
\alias{true_global_rt}
\title{Example hospital admissions dataset for evaluation using  exp. corr. func.
Global R(t) estimate dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 127 rows and 4 columns.
}
\source{
vignette_data.R
}
\usage{
true_global_rt
}
\description{
A dataset containing the the simulated unadjusted daily R(t) estimate
as well as the realized R(t) estimate for the overall "global" population.
This realized R(t) is generated by summing the incident infections from
each of the subpopulations (the populations in each wastewater catchment
area plus an additional one representing the remainder of the population).
The R(t) is then back-calculated by dividing the total incident infection
time seres by the convolution of the infection time series and the
generation interval.
}
\details{
This data is generated via the default values in the
\code{generate_simulated_data()} function.

The variables are as follows:
\describe{
\item{unadj_rt}{The daily global unadjusted R(t). The log of the
unadjusted R(t) is the central value around which the subpopulation
R(t) values are drawn from each week.}
\item{realized_rt}{The daily value of the realized global R(t). This is
calculated by summing up the subpopulation level incident infections and
then dividing by the convolution of the total incident infections and the
generation interval to get the R(t) value that corresponds to the
global incident infections}
\item{t}{The time index in days}
\item{date}{Date the hospital admissions occurred, formatte din ISO8601
standatds as YYYY-MM-DD}
}
}
\keyword{datasets}
