% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_component_fwd_sim.R
\name{truncate_for_latency}
\alias{truncate_for_latency}
\title{Truncate the predicted wastewater concentrations based on the
lab site reporting latency and the observed time and horizon time}
\usage{
truncate_for_latency(
  log_conc_lab_site,
  n_lab_sites,
  ot,
  ht,
  nt,
  lab_site_reporting_latency
)
}
\arguments{
\item{log_conc_lab_site}{The matrix of n_lab_sites by n time points
indicating the underlying expected  observed concentrations}

\item{n_lab_sites}{Integer indicating the number of unique lab-site
combinations}

\item{ot}{integer indicating the number of days we will have observed data
for in the calibration period}

\item{ht}{integer indicating the time after the last observed time to
the end of the forecast time}

\item{nt}{integer indicating the time after the last observed epi indicator
and before the forecast date, of which there can still be wastewater
observations}

\item{lab_site_reporting_latency}{vector indicating the number of days
from the forecast date of the last possible observation}
}
\value{
A sparse matrix of \code{n_lab_sites} rows and \code{ot} + \code{ht} columns of
but with NAs for when observations are not measured/reported.
}
\description{
Truncate the predicted wastewater concentrations based on the
lab site reporting latency and the observed time and horizon time
}
