% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_count_data}
\alias{validate_count_data}
\title{Validate user-provided count data}
\usage{
validate_count_data(
  count_data,
  count_col_name,
  pop_size_col_name,
  call = rlang::caller_env()
)
}
\arguments{
\item{count_data}{tibble containing the input count data}

\item{count_col_name}{string indicating the name of the column containing
the count data}

\item{pop_size_col_name}{string indicating the name of the column containing
the population size of the count catchment area}

\item{call}{Calling environment to be passed to \code{\link[cli:cli_abort]{cli::cli_abort()}} for
traceback.}
}
\value{
NULL, invisibly
}
\description{
Validate user-provided count data
}
