% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ww_data}
\alias{ww_data}
\title{Example wastewater dataset.}
\format{
\subsection{ww_data}{

A tibble with 102 rows and 6 columns
\describe{
\item{date}{Sample collection date, formatted in ISO8601 standards as
YYYY-MM-DD}
\item{site}{The wastewater treatment plant where the sample was collected}
\item{lab}{The lab where the sample was processed}
\item{genome_copies_per_ml}{The wastewater concentration measured on the
date specified, collected in the site specified, and processed in the lab
specified. The default parameters assume that this quantity is reported
as the genome copies per mL, on a natural scale.}
\item{lod}{The limit of detection in the site and lab on a particular day
of the quantification device (e.g. PCR). This is also by default reported
in terms of the genome copies per mL.}
\item{site_pop}{The population size of the wastewater catchment area
represented by the site variable}
}
}
}
\source{
vignette_data.R
}
\usage{
ww_data
}
\description{
A dataset containing the simulated wastewater concentrations
(labeled here as \code{genome_copies_per_ml}) by sample collection date (\code{date}),
the site where the sample was collected (\code{site}) and the lab where the
samples were processed (\code{lab}). Additional columns that are required
attributes needed for the model are the limit of detection for that lab on
each day (labeled here as \code{lod}) and the population size of the wastewater
catchment area represented by the wastewater concentrations in each \code{site}.
}
\details{
This data is generated via the default values in the
\code{generate_simulated_data()} function. They represent the bare minumum
required fields needed to pass to the model, and we recommend that users
try to format their own data to match this format.

The variables are as follows:
}
\keyword{datasets}
