% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{assert_equivalent_indexing}
\alias{assert_equivalent_indexing}
\title{Assert that two tibbles of date and time mapping align}
\usage{
assert_equivalent_indexing(
  full_data,
  subsetted_data,
  call = rlang::caller_env(),
  add_err_msg = ""
)
}
\arguments{
\item{full_data}{a tibble containing the columns \code{date} (with IS08 dates)
and \code{t} (integers of time in days) that the other tibble will be joined to}

\item{subsetted_data}{a tibble containing the columns \code{date} (with IS08 dates)
and \code{t} (integers of time in days) that we indexed based on the \code{full_data}}

\item{call}{Calling environment to be passed to \code{\link[cli:cli_abort]{cli::cli_abort()}} for
traceback.}

\item{add_err_msg}{add_err_msg string containing an additional error message,
default is the empty string (\code{""})}
}
\value{
NULL invisible
}
\description{
Assert that two tibbles of date and time mapping align
}
