% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{assert_overlap_dates}
\alias{assert_overlap_dates}
\title{Assert that two vectors of dates overlap}
\usage{
assert_overlap_dates(
  dates1,
  dates2,
  max_date,
  call = rlang::caller_env(),
  add_err_msg = ""
)
}
\arguments{
\item{dates1}{the vector of dates to check, must of date type}

\item{dates2}{the vector of dates to compare to, must be of date type}

\item{max_date}{the maximum date the testing dates should be}

\item{call}{Calling environment to be passed to \code{\link[cli:cli_abort]{cli::cli_abort()}} for
traceback.}

\item{add_err_msg}{add_err_msg string containing an additional error message,
default is the empty string (\code{""})}
}
\value{
NULL invisible
}
\description{
Assert that two vectors of dates overlap
}
