% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_component_fwd_sim.R
\name{calc_rt}
\alias{calc_rt}
\title{Back- calculate R(t) from incident infections and the generation interval}
\usage{
calc_rt(new_i, convolve_fxn, generation_interval, uot, tot_time)
}
\arguments{
\item{new_i}{vector of numerics that spans the length of \code{tot_time},
representing the new incident infections per day}

\item{convolve_fxn}{function used to convolve infections with delay pmf
This will typically take \code{convolve_dot_product()} from the
\code{wwinference.stan} model}

\item{generation_interval}{vector of simplex describing the probability of
each time from infection to onwards transmission}

\item{uot}{integer indicating the days for exponential growth initialization
to occur (referred to as unobserved time)}

\item{tot_time}{integer indicating the total time we have incident
infections for}
}
\value{
a numeric vector of length(\code{tot_time} - \code{uot}) that represents
the effective reproductive number
}
\description{
Note that the forward renewal process is not a simple convolution of
incident infections and the generation interval -- because it assumes that
the generation interval being passed in is indexed starting at day 1, and
that on any particular index day there is no contribution from individuals
infected on that day. This is a reasonable assumption, but to align the
implementation of the forward process with our backward process, we have to
add a 0 density to day 0 of the passed in generation interval.
}
