% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_elements_non_neg}
\alias{check_elements_non_neg}
\title{Check that all elements of a vector are non-negative}
\usage{
check_elements_non_neg(x, arg = "x", call = rlang::caller_env())
}
\arguments{
\item{x}{vector of arguments to check for negativity}

\item{call}{Calling environment to be passed to the type checker}
}
\value{
NULL, invisibly
}
\description{
Check that all elements of a vector are non-negative
}
