% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkers.R
\name{check_required_ww_inputs}
\alias{check_required_ww_inputs}
\title{Check that the input wastewater data contains all the required column names}
\usage{
check_required_ww_inputs(
  ww_data,
  conc_col_name,
  lod_col_name,
  call = rlang::caller_env()
)
}
\arguments{
\item{ww_data}{tibble containing the input wastewater data}

\item{conc_col_name}{string indicating the name of the column containing
the concentration measurements in the wastewater data}

\item{lod_col_name}{string indicating the name of the column containing
the concentration measurements in the wastewater data}

\item{call}{Calling environment to be passed to \code{\link[cli:cli_abort]{cli::cli_abort()}} for
traceback.}
}
\value{
NULL, invisibly
}
\description{
Check that the input wastewater data contains all the required column names
}
