% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_component_fwd_sim.R
\name{create_site_lab_map}
\alias{create_site_lab_map}
\title{Create a mapping of sites, labs, and population sizes in each site}
\usage{
create_site_lab_map(site, lab, ww_pop_sites)
}
\arguments{
\item{site}{vector of integers or characters uniquely identifying a site,
or the population in the wastewater catchment area}

\item{lab}{vector of integer or character uniquely identifying a lab where
a sample was processed}

\item{ww_pop_sites}{vector of integers indicating the population size in
each site}
}
\value{
a tibble mapping sites, labs and population sizes
}
\description{
Create a mapping of sites, labs, and population sizes in each site
}
