% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_component_fwd_sim.R
\name{format_ww_data}
\alias{format_ww_data}
\title{Format the wastewater data as a tidy data frame}
\usage{
format_ww_data(
  log_obs_conc_lab_site,
  ot,
  ht,
  date_df,
  site_lab_map,
  lod_lab_site
)
}
\arguments{
\item{log_obs_conc_lab_site}{matrix of numeric values where rows are the
site-lab combinations and columns are the observed time points}

\item{ot}{integer indicating the number of days we will have observed data
for in the calibration period}

\item{ht}{integer indicating the time after the last observed epi indicator
and before the forecast date, of which there can still be wastewater
observations}

\item{date_df}{tibble of columns \code{date} and \code{t} that map time in days to
dates}

\item{site_lab_map}{tibble mapping sites, labs, lab site indices, and the
population size of the site}

\item{lod_lab_site}{vector of numerics indicating the LOD in each lab and
site combination}
}
\value{
a tidy dataframe containing the observed wastewater concentrations
in each site and lab at each time point
}
\description{
Format the wastewater data as a tidy data frame
}
\seealso{
Other Model components:
\code{\link{calc_rt}()},
\code{\link{create_site_lab_map}()},
\code{\link{downsample_ww_obs}()},
\code{\link{format_hosp_data}()},
\code{\link{get_global_rt}()},
\code{\link{get_pred_obs_conc}()},
\code{\link{get_pred_subpop_gen_per_n}()},
\code{\link{get_time_varying_daily_ihr}()},
\code{\link{subpop_inf_process}()},
\code{\link{subpop_rt_process}()}
}
\concept{Model components}
