% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_draws.R
\name{get_draws}
\alias{get_draws}
\alias{get_draws_df}
\alias{get_draws.wwinference_fit}
\alias{get_draws.default}
\alias{get_draws.data.frame}
\alias{plot.wwinference_fit_draws}
\title{Postprocess to generate a draws dataframe}
\usage{
get_draws(x, ..., what = "all")

get_draws_df(x, ...)

\method{get_draws}{wwinference_fit}(x, ..., what = "all")

\method{get_draws}{default}(x, ..., what = "all")

\method{get_draws}{data.frame}(
  x,
  count_data,
  date_time_spine,
  site_subpop_spine,
  lab_site_subpop_spine,
  stan_data_list,
  fit_obj,
  ...,
  what = "all"
)

\method{plot}{wwinference_fit_draws}(x, y = NULL, what, ...)
}
\arguments{
\item{x}{An object of class \code{get_draws}.}

\item{...}{additional arguments}

\item{what}{Character vector. Specifies the variables to extract from the
draws. It could be any from \code{"all"} \code{"predicted_counts"}, \code{"predicted_ww"},
\code{"global_rt"}, or \code{"subpop_rt"}. When \code{what = "all"} (the default),
the function will extract all four variables.}

\item{count_data}{A dataframe of the preprocessed daily count data (e.g.
hospital admissions) from the "global" population}

\item{date_time_spine}{tibble mapping dates to time in days}

\item{site_subpop_spine}{tibble mapping sites to subpopulations}

\item{lab_site_subpop_spine}{tibble mapping lab-sites to subpopulations}

\item{stan_data_list}{A list containing all the data passed to stan for
fitting the model}

\item{fit_obj}{a CmdStan object that is the output of fitting the model to
\code{x} and \code{count_data}}

\item{y}{Ignored in the the case of \code{plot}.}
}
\value{
A tibble containing the full set of posterior draws of the
estimated, nowcasted, and forecasted: counts, site-level wastewater
concentrations, "global"(e.g. state) R(t) estimate, and the  "local" (site +
the one auxiliary subpopulation) R(t) estimates. In the instance where there
are observations, the data will be joined to each draw of the predicted
observation to facilitate plotting.
}
\description{
This function takes in the two input data sources, the CmdStan fit object,
and the 3 relevant mappings from stan indices to the real data, in order
to generate a dataframe containing the posterior draws of the counts (e.g.
hospital admissions), the wastewater concentration values, the "global" R(t),
and the "local" R(t) estimates + the critical metadata in the data.
This funtion has a default method that takes the two sets of input data,
the last of stan arguments, and the CmdStan fitting object, as well as an S3
method for objects of class 'wwinference_fit'

This method overloads the generic \code{get_draws} function specifically
for objects of type 'wwinference_fit'.
}
\details{
The function \code{get_draws_df()} has been deprecated in favor of \code{get_draws()}.

The plot method for \code{wwinference_fit_draws} is a wrapper of
\code{get_plot_forecasted_counts}, \code{get_plot_ww_conc}, \code{get_plot_global_rt},
and \code{get_plot_subpop_rt}. Depending on the value of \code{what}, the function
will call the appropriate method.
}
