% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_draws_df.R
\name{get_draws_df}
\alias{get_draws_df}
\alias{get_draws_df.wwinference_fit}
\alias{get_draws_df.default}
\title{Postprocess to generate a draws dataframe}
\usage{
get_draws_df(ww_output, ...)

\method{get_draws_df}{wwinference_fit}(ww_output, ...)

\method{get_draws_df}{default}(ww_data, count_data, stan_args, fit_obj)
}
\arguments{
\item{ww_output}{an object of the \code{wwinference_fit} class}

\item{...}{additional arguments}

\item{ww_data}{A dataframe of the preprocessed wastewater concentration data
used to fit the model}

\item{count_data}{A dataframe of the preprocessed daily count data (e.g.
hospital admissions) from the "global" population}

\item{stan_args}{A list containing all the data passed to stan for fitting
the model}

\item{fit_obj}{a CmdStan object that is the output of fitting the model to
the \code{ww_data} and \code{count_data}}
}
\value{
A tibble containing the full set of posterior draws of the
estimated, nowcasted, and forecasted: counts, site-level wastewater
concentrations, "global"(e.g. state) R(t) estimate, and the  "local" (site +
the one auxiliary subpopulation) R(t) estimates. In the instance where there
are observations, the data will be joined to each draw of the predicted
observation to facilitate plotting.
}
\description{
This function takes in the two input data sources, the CmdStan fit object,
and the 3 relevant mappings from stan indices to the real data, in order
to generate a dataframe containing the posterior draws of the counts (e.g.
hospital admissions), the wastewater concentration values, the "global" R(t),
and the "local" R(t) estimates + the critical metadata in the data
}
