% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_draws_df.R
\name{get_draws_df}
\alias{get_draws_df}
\title{Postprocess to generate a draws dataframe}
\usage{
get_draws_df(
  ww_data,
  count_data,
  fit_obj,
  date_time_spine,
  lab_site_spine,
  subpop_spine
)
}
\arguments{
\item{ww_data}{A dataframe of the preprocessed wastewater concentration data
used to fit the model}

\item{count_data}{A dataframe of the preprocessed daily count data (e.g.
hospital admissions) from the "global" population}

\item{fit_obj}{a CmdStan object that is the output of fitting the model to
the \code{ww_data} and \code{count_data}}

\item{date_time_spine}{A tibble mapping the time index in stan (observed +
nowcast + forecast) to real dates}

\item{lab_site_spine}{A tibble mapping the site-lab index in stan to the
corresponding site, lab, and site population}

\item{subpop_spine}{A tibble mapping the site index in stan to the
corresponding subpopulation (either a site or the auxiliary site we add to
represent the rest of the population)}
}
\value{
A tibble containing the full set of posterior draws of the
estimated, nowcasted, and forecasted: counts, site-level wastewater
concentrations, "global"(e.g. state) R(t) estimate, and the  "local" (site +
the one auxiliary subpopulation) R(t) estimates. In the instance where there
are observations, the data will be joined to each draw of the predicted
observation to facilitate plotting.
}
\description{
This function takes in the two input data sources, the CmdStan fit object,
and the 3 relevant mappings from stan indices to the real data, in order
to generate a dataframe containing the posterior draws of the counts (e.g.
hospital admissions), the wastewater concentration values, the "global" R(t),
and the "local" R(t) estimates + the critical metadata in the data
}
