% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_component_fwd_sim.R
\name{get_global_rt}
\alias{get_global_rt}
\title{Function to generate a global weekly R(t) estimate with added noise}
\usage{
get_global_rt(r_in_weeks, n_weeks, global_rt_sd)
}
\arguments{
\item{r_in_weeks}{The mean R(t) for each week of the simulation, can be
longer than the 7*n_weeks}

\item{n_weeks}{The number of weeks that we want to simulate the data for}

\item{global_rt_sd}{The variation in the R(t) estimate to add
intrinsic variability to the infection dynamics}
}
\value{
a weekly R(t) estimate with added noise
}
\description{
Function to generate a global weekly R(t) estimate with added noise
}
\seealso{
Other Model components:
\code{\link{calc_rt}()},
\code{\link{create_site_lab_map}()},
\code{\link{downsample_ww_obs}()},
\code{\link{format_hosp_data}()},
\code{\link{format_ww_data}()},
\code{\link{get_pred_obs_conc}()},
\code{\link{get_pred_subpop_gen_per_n}()},
\code{\link{get_time_varying_daily_ihr}()},
\code{\link{subpop_inf_process}()},
\code{\link{subpop_rt_process}()}
}
\concept{Model components}
