% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stan_data.R
\name{get_input_count_data_for_stan}
\alias{get_input_count_data_for_stan}
\title{Get the input count data to pass directly to stan}
\usage{
get_input_count_data_for_stan(preprocessed_count_data, calibration_time)
}
\arguments{
\item{preprocessed_count_data}{a dataframe with the input count data, must
have the following columns: date, count, total_pop}

\item{calibration_time}{integer indicating the max duration in days that
the model is calibrated to the count data for}
}
\value{
datatframe of count data passed to stan
}
\description{
Get the input count data to pass directly to stan
}
