% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wwinference.R
\name{get_mcmc_options}
\alias{get_mcmc_options}
\title{Get MCMC options}
\usage{
get_mcmc_options(
  iter_warmup = 750,
  iter_sampling = 500,
  parallel_chains = 4,
  seed = NULL,
  adapt_delta = 0.95,
  max_treedepth = 12
)
}
\arguments{
\item{iter_warmup}{integer indicating the number of warm-up iterations,
default is \code{750}.}

\item{iter_sampling}{integer indicating the number of sampling iterations,
default is \code{500}.}

\item{parallel_chains}{integer indicating the number of chains to run
in parallel, default is \code{4}.}

\item{seed}{set of integers indicating the random seed of the Stan sampler,
default is \code{NULL}.}

\item{adapt_delta}{float between 0 and 1 indicating the average acceptance
probability, default is \code{0.95}.}

\item{max_treedepth}{integer indicating the maximum tree depth of the
sampler, default is 12.}
}
\value{
A list of MCMC settings with the values given by the function.
arguments
}
\description{
This function returns a list of MCMC settings to pass to the
\code{cmdstanr::sample()} function to fit the model. The default settings are
specified for production-level runs, consider adjusting to optimize
for speed while iterating. All input arguments to \code{cmdstanr::sample()}
are configurable by the user. See \code{cmdstanr::sample()} documentation
https://mc-stan.org/cmdstanr/reference/model-method-sample.html for a full
description of each argument.
}
