% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wwinference.R
\name{get_mcmc_options}
\alias{get_mcmc_options}
\title{Get MCMC options}
\usage{
get_mcmc_options(
  iter_warmup = 750,
  iter_sampling = 500,
  n_chains = 4,
  seed = 123,
  adapt_delta = 0.95,
  max_treedepth = 12,
  compute_likelihood = 1
)
}
\arguments{
\item{iter_warmup}{integer indicating the number of warm-up iterations,
default is \code{750}}

\item{iter_sampling}{integer indicating the number of sampling iterations,
default is \code{500}}

\item{n_chains}{integer indicating the number of MCMC chains to run, default
is \code{4}}

\item{seed}{set of integers indicating the random seed of the stan sampler,
default is \code{123}}

\item{adapt_delta}{float between 0 and 1 indicating the average acceptance
probability, default is \code{0.95}}

\item{max_treedepth}{integer indicating the maximum tree depth of the
sampler, default is 12}

\item{compute_likelihood}{integer indicating whether or not to compute the
likelihood using the data, default is \code{1} which will fit the model to the
data. If set to 0, the model will sample from the prior only}
}
\value{
a list of mcmc settings with the values given by the  function
arguments
}
\description{
This function returns a list of MCMC settings to pass to the
\code{cmdstanr::sample()} function to fit the model. The default settings are
specified for production-level runs, consider adjusting to optimize
for speed while iterating.
}
\examples{
mcmc_settings <- get_mcmc_options()
}
