% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wwinference.R
\name{get_model_spec}
\alias{get_model_spec}
\title{Get model specifications}
\usage{
get_model_spec(
  generation_interval = wwinference::generation_interval,
  inf_to_count_delay = wwinference::inf_to_hosp,
  infection_feedback_pmf = wwinference::generation_interval,
  include_ww = TRUE,
  params = get_params(system.file("extdata", "example_params.toml", package =
    "wwinference"))
)
}
\arguments{
\item{generation_interval}{vector of a simplex (must sum to 1) describing
the daily probability of onwards transmission, default is package data
provided for the COVID-19 generation interval post-Omicron}

\item{inf_to_count_delay}{vector of a simplex (must sum to 1) describing the
daily probability of transitioning from infection to whatever the count
variable is, e.g. hospital admissions or cases. Default corresonds to the
delay distribution from COVID-19 infection to hospital admission}

\item{infection_feedback_pmf}{vector of a simplex (must sum to 1) describing
the delay from incident infection to feedback in the transmission dynamics.
The default is the COVID-19 generation interval}

\item{include_ww}{Boolean indicating whether or not to include the wastewater
data into the model, default is \code{TRUE} which will get passed to stan and
tell the model to evaluate the likelihood with the wastewater data}

\item{params}{a list of parameters corresponding to model
priors and disease/data specific parameters. Default is for COVID-19 hospital
admissions and viral concentrations in wastewater}
}
\value{
a list of model specs to be passed to the \code{get_stan_data()} function
}
\description{
This function returns a nested list containing the model specifications
in the function arguments. All defaults are set for the case of fitting a
post-omicron COVID-19 model with joint inference of hospital admissions
and data on wastewater viral concentrations
}
\examples{
model_spec_list <- get_model_spec()
}
