% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stan_data.R
\name{get_stan_data}
\alias{get_stan_data}
\title{Get stan data for ww + hosp model}
\usage{
get_stan_data(
  input_count_data,
  input_ww_data,
  forecast_date,
  forecast_horizon,
  calibration_time,
  generation_interval,
  inf_to_count_delay,
  infection_feedback_pmf,
  params,
  compute_likelihood = 1
)
}
\arguments{
\item{input_count_data}{a dataframe with the input count data, must have
the following columns: date, count, total_pop}

\item{input_ww_data}{a dataframe with the input wastewater data with no gaps,
must have the following columns: date, site, lab, genome_copies_per_ml,
site_pop, below_lod, and if removing outliers, flag_as_ww_outlier}

\item{forecast_date}{string indicating the forecast date in ISO8601
convention e.g. YYYY-MM-DD}

\item{forecast_horizon}{integer indicating the number of days to make a
forecast for}

\item{calibration_time}{integer indicating the max duration in days that
the model is calibrated to the count data for}

\item{generation_interval}{a vector with a zero-truncated normalized pmf of
the generation interval}

\item{inf_to_count_delay}{a vector with a normalized pmf of the delay from
infection to counts}

\item{infection_feedback_pmf}{a vector with a normalized pmf dictating the
delay of infection feedback}

\item{params}{a dataframe of parameter names and numeric values}

\item{compute_likelihood}{indicator variable telling stan whether or not to
compute the likelihood, default = \code{1}}
}
\value{
a list of named variables to pass to stan
}
\description{
Get stan data for ww + hosp model
}
