% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_component_fwd_sim.R
\name{get_time_varying_daily_ihr}
\alias{get_time_varying_daily_ihr}
\title{Get time varying IHR}
\usage{
get_time_varying_daily_ihr(p_hosp_mean, uot, ot, ht, tot_weeks, p_hosp_w_sd_sd)
}
\arguments{
\item{p_hosp_mean}{mean probability of hospital admission given infection}

\item{uot}{integer indicating the days for exponential growth initialization
to occur (referred to as unobserved time)}

\item{ot}{integer indicating the number of days we will have observed data
for in the calibration period}

\item{ht}{integer indicating the time after the last observed time to
forecast}

\item{tot_weeks}{integer indicating the total number of weeks including the
\code{uot}}

\item{p_hosp_w_sd_sd}{the standard deviation in logit scale of the
IHR}
}
\value{
a vector of daily IHR values
}
\description{
Get time varying IHR
}
\seealso{
Other Model components:
\code{\link{calc_rt}()},
\code{\link{create_site_lab_map}()},
\code{\link{downsample_ww_obs}()},
\code{\link{format_hosp_data}()},
\code{\link{format_ww_data}()},
\code{\link{get_global_rt}()},
\code{\link{get_pred_obs_conc}()},
\code{\link{get_pred_subpop_gen_per_n}()},
\code{\link{subpop_inf_process}()},
\code{\link{subpop_rt_process}()}
}
\concept{Model components}
