% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_diagnostics.R
\name{parameter_diagnostics}
\alias{parameter_diagnostics}
\title{Method for printing the CmdStan parameter diagnostics for a
wwinference_fit_object}
\usage{
parameter_diagnostics(ww_fit, ...)
}
\arguments{
\item{ww_fit}{An object of class wwinference_fit}
}
\description{
Method for printing the CmdStan parameter diagnostics for a
wwinference_fit_object
}
\seealso{
Other diagnostics:
\code{\link{diagnostic_summary}()},
\code{\link{get_model_diagnostic_flags}()}
}
\concept{diagnostics}
