% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocess_hosp_data}
\alias{preprocess_hosp_data}
\title{Pre-process hospital admissions data, converting column names to those
that \code{\link[=get_stan_data]{get_stan_data()}} expects.}
\usage{
preprocess_hosp_data(
  hosp_data,
  count_col_name = "daily_hosp_admits",
  pop_size_col_name = "state_pop"
)
}
\arguments{
\item{hosp_data}{dataframe containing the following columns: date,
a count column, and a population size column}

\item{count_col_name}{name of the column containing the epidemiological
indicator, default is \code{daily_hosp_admits}}

\item{pop_size_col_name}{name of the column containing the population size
of that the counts are coming from, default is \code{state_pop}}
}
\value{
a dataframe containing the hospital admissions data renamed to
have the following columns \code{date}, \code{count}, and \code{total_pop}
}
\description{
Pre-process hospital admissions data, converting column names to those
that \code{\link[=get_stan_data]{get_stan_data()}} expects.
}
\examples{
hosp_data <- tibble::tibble(
  date = c("2023-11-01", "2023-11-02"),
  daily_admits = c(10, 20),
  state_pop = c(1e6, 1e6)
)
hosp_data_preprocessed <- preprocess_hospdata(
  hosp_data,
  "daily_admits",
  "state_pop"
)
}
