% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{preprocess_ww_data}
\alias{preprocess_ww_data}
\title{Pre-process wastewater input data, adding needed indices and flagging
potential outliers}
\usage{
preprocess_ww_data(
  ww_data,
  conc_col_name = "genome_copies_per_ml",
  lod_col_name = "lod"
)
}
\arguments{
\item{ww_data}{dataframe containing the following columns: site, lab,
date, a column for concentration, and lod}

\item{conc_col_name}{string indicating the name of the column containing
the concentration measurements in the wastewater data, default is
\code{genome_copies_per_ml}}

\item{lod_col_name}{string indicating the name of the column containing
the concentration measurements in the wastewater data, default is
\code{genome_copies_per_ml}}
}
\value{
a dataframe containing the same columns as ww_data except
the \code{conc_col_name} will be replaced with \code{genome_copies_per_ml} and
the \code{lod_col_name} will be replaced with \code{lod} plus the following
additional columns needed for the stan model:
lab_site_index, site_index, flag_as_ww_outlier, lab_site_name,
forecast_date
}
\description{
Pre-process wastewater input data, adding needed indices and flagging
potential outliers
}
\examples{
ww_data <- tibble::tibble(
  date = rep(c("2023-11-01", "2023-11-02"), 2),
  site = c(rep(1, 2), rep(2, 2)),
  lab = c(1, 1, 1, 1),
  conc = c(345.2, 784.1, 401.5, 681.8),
  lod = c(20, 20, 15, 15)
)
ww_data_preprocessed <- preprocess_ww_data(ww_data,
                                           conc_col_name = "conc",
                                           lod_col_name = "lod"
                                           )
}
