% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subpop_hosp_data}
\alias{subpop_hosp_data}
\title{\item{date}{Date the hospital admissions occurred, formatted in ISO8601
standards as YYYY-MM-DD}
\item{subpop_name}{A string indicating the subpopulation the hospital
admissiosn corresponds to. This is either a wastewater site, or the
remainder of the population}
\item{daily_hosp_admits}{The number of individuals admitted to the
hospital on that date, available as of the forecast date}
\item{subpop_pop}{The number of people contributing to the daily hospital
admissions in each subpopulation}
}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 450 rows and 4 columns.
}
\source{
vignette_data.R
}
\usage{
subpop_hosp_data
}
\description{
\item{date}{Date the hospital admissions occurred, formatted in ISO8601
standards as YYYY-MM-DD}
\item{subpop_name}{A string indicating the subpopulation the hospital
admissiosn corresponds to. This is either a wastewater site, or the
remainder of the population}
\item{daily_hosp_admits}{The number of individuals admitted to the
hospital on that date, available as of the forecast date}
\item{subpop_pop}{The number of people contributing to the daily hospital
admissions in each subpopulation}
}
}
\keyword{datasets}
