% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subpop_hosp_data}
\alias{subpop_hosp_data}
\title{Example subpopulation level hospital admissions dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 450 rows and 4 columns.
}
\source{
vignette_data.R
}
\usage{
subpop_hosp_data
}
\description{
A dataset containing the simulated daily hospital admissions
(labeled here as \code{daily_hosp_admits}) by date of admission (\code{date}) in
each subpopulation.
Additional columns that are required are the population size of the
population contributing to the hospital admissions. It is assumed that
the wastewater sites are subsets of this larger population, which
is in the package data assumed to be from a hypothetical US state.
The data generated are daily hospital admissions but they could be any other
epidemiological count dataset e.g. cases. This data should only contain
hospital admissions that would have been available as of the date that
the forecast was made. We recommend that users try to format their data
to match this format.
}
\details{
This data is generated via the default values in the
\code{generate_simulated_data()} function. They represent the bare minimumum
required fields needed to pass to the model, and we recommend that users
try to format their own data to match this format.

The variables are as follows:
\describe{
\item{date}{Date the hospital admissions occurred, formatted in ISO8601
standards as YYYY-MM-DD}
\item{subpop_name}{A string indicating the subpopulation the hospital
admissiosn corresponds to. This is either a wastewater site, or the
remainder of the population}
\item{daily_hosp_admits}{The number of individuals admitted to the
hospital on that date, available as of the forecast date}
\item{subpop_pop}{The number of people contributing to the daily hospital
admissions in each subpopulation}
}
}
\keyword{datasets}
