% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_component_fwd_sim.R
\name{subpop_rt_process}
\alias{subpop_rt_process}
\title{Get subpopulation level R(t) estimates assuming time and space independence}
\usage{
subpop_rt_process(n_sites, r_weeks, subpop_level_rt_variation)
}
\arguments{
\item{n_sites}{integer indicating the number of wastewater sites, assumes
we will always have n_sites + 1 subpopulations}

\item{r_weeks}{The "global" R(t) in weeks}

\item{subpop_level_rt_variation}{The standard deviation of the Gaussian
to generate deviation at the site level in the R(t) estimate}
}
\value{
A n_sites+1 by n_weeks matrix containing the subpopulation R(t)s
}
\description{
Get subpopulation level R(t) estimates assuming time and space independence
}
