% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_diagnostics.R
\name{summary_diagnostics}
\alias{summary_diagnostics}
\title{Method for printing the CmdStan summary diagnostics for
wwinference_fit_object}
\usage{
summary_diagnostics(ww_fit, ...)
}
\arguments{
\item{ww_fit}{An object of class wwinference_fit}

\item{...}{additional arguments}
}
\description{
Method for printing the CmdStan summary diagnostics for
wwinference_fit_object
}
\seealso{
Other diagnostics:
\code{\link{get_model_diagnostic_flags}()},
\code{\link{parameter_diagnostics}()},
\code{\link{wwinference}()}
}
\concept{diagnostics}
