% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_both_datasets}
\alias{validate_both_datasets}
\title{Validate that both count data and wastewater data are coherent and
compatible with one another and the the user-specified parameters}
\usage{
validate_both_datasets(
  input_count_data,
  input_ww_data,
  calibration_time,
  forecast_date
)
}
\arguments{
\item{input_count_data}{tibble containing the input count data that has
been filtered and is ready to be passed into stan}

\item{input_ww_data}{tibble containing the input wastewater data that has
been filtered and is ready to be passed into stan}

\item{calibration_time}{integer indicating the calibration time}

\item{forecast_date}{IS08 formatted date indicating the forecast date}
}
\value{
NULL, invisibly
}
\description{
Validate that both count data and wastewater data are coherent and
compatible with one another and the the user-specified parameters
}
