% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_pmf}
\alias{validate_pmf}
\title{Validate that the pmf being passed to stan}
\usage{
validate_pmf(
  pmf,
  calibration_time,
  count_data,
  arg = "x",
  call = rlang::caller_env()
)
}
\arguments{
\item{pmf}{simplex vector describing a probabilty of an event ocurring on
each day}

\item{calibration_time}{integer indicating the calibration time}

\item{arg}{name of the argument supplying the object}

\item{call}{The calling environment to be reflected in the error message}

\item{tibble}{containing the input count data ready to be passed to stan}
}
\value{
NULL, invisibly
}
\description{
Validate that the pmf being passed to stan
}
