% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wwinference.R
\name{wwinference}
\alias{wwinference}
\title{Fit the joint inference of wastewater and count data}
\usage{
wwinference(
  ww_data,
  count_data,
  model_spec = get_model_spec(forecast_date = "2023-12-06"),
  mcmc_options = get_mcmc_options(),
  compiled_model = compile_model()
)
}
\arguments{
\item{ww_data}{}

\item{count_data}{}

\item{model_spec}{}

\item{mcmc_options}{}

\item{compiled_model}{}
}
\description{
Fit the joint inference of wastewater and count data
}
