% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wwinference.R
\name{wwinference}
\alias{wwinference}
\title{Joint inference of count data (e.g. cases/admissions) and wastewater
data}
\usage{
wwinference(
  ww_data,
  count_data,
  forecast_date = NULL,
  calibration_time = 90,
  forecast_horizon = 28,
  model_spec = get_model_spec(),
  mcmc_options = get_mcmc_options(),
  generate_initial_values = TRUE,
  compiled_model = compile_model(),
  dist_matrix = NULL,
  corr_func = 0
)
}
\arguments{
\item{ww_data}{A dataframe containing the pre-processed, site-level
wastewater concentration data for a model run. The dataframe must contain
the following columns: \code{date}, \code{site}, \code{lab}, \code{genome_copies_per_ml},
\code{lab_site_index}, \code{lod}, \code{below_lod}, \code{site_pop} \code{exclude}}

\item{count_data}{A dataframe containing the pre-procssed, "global" (e.g.
state) daily count data, pertaining to the number of events that are being
counted on that day, e.g. number of daily cases or daily hospital admissions.
Must contain the following columns: \code{date}, \code{count} , \code{total_pop}}

\item{forecast_date}{a character string in ISO8601 format (YYYY-MM-DD)
indicating the date that the forecast is to be made. Default is NULL}

\item{calibration_time}{integer indicating the number of days to calibrate
the model for, default is \code{90}}

\item{forecast_horizon}{integer indicating the number of days, including the
forecast date, to produce forecasts for, default is \code{28}}

\item{model_spec}{The model specification parameters as defined using
\code{get_model_spec()}. The default here pertains to the \code{forecast_date} in the
example data provided by the package, but this should be specified by the
user based on the date they are producing a forecast}

\item{mcmc_options}{The MCMC parameters as defined using
\code{get_mcmc_options()}.}

\item{generate_initial_values}{Boolean indicating whether or not to specify
the initialization of the sampler, default is \code{TRUE}, meaning that
initialization lists will be generated and passed as the \code{init} argument
to the model object \code{\link[cmdstanr:model-method-sample]{$sample()}} call.
function}

\item{compiled_model}{The pre-compiled model as defined using
\code{compile_model()}}

\item{dist_matrix}{Distance matrix, n_sites x n_sites, passed to a
distance-based correlation function for epsilon. If NULL, use an independence
correlation function, for current implementation (i.e. all sites' epsilon
values are independent and identically distributed) .}

\item{corr_func}{Integer variable to define the type of correlation
function used : 0-iid, 1-exponential, 2-lkj.}
}
\value{
A nested list of the following items, intended to allow the user to
quickly and easily plot results from their inference, while also being able
to have the full user functionality of running the model themselves in stan
by providing the raw model object and diagnostics. If the model runs, this
function will return:
\code{draws_df}: A tibble containing the full set of posterior draws of the
estimated, nowcasted, and forecasted: counts, site-level wastewater
concentrations, "global"(e.g. state) R(t) estimate, and the  "local" (site +
the one auxiliary subpopulation) R(t) estimates. In the instance where there
are observations, the data will be joined to each draw of the predicted
observation to facilitate plotting.
\code{raw_fit_obj}: The CmdStan object that is returned from the call to
\code{cmdstanr::sample()}. Can be used to access draws, summary, diagnostics, etc.
\code{date_time_spine}: Mapping from time in stan to dates
\code{lab_site_spine}: Mapping from lab_site_index in stan to lab and site
\code{subpop_spine}: Mapping from site index in stan to site

If the model fails to run, a list containing the follow will be returned:
\code{error}: the error message provided from stan, indicating why the model
failed to run. Note, the model might still run and produce draws even if it
has major model issues. We recommend the user always run the
\code{check_diagnostics()} function on the \code{diagnostic_summary} as part of any
pipeline to ensure model convergence.
}
\description{
Provides a user friendly interface around package functionality
to produce estimates, nowcasts, and forecasts pertaining to user-specified
delay distributions, set parameters, and priors that can be modified to
handledifferent types of "global" count data and "local" wastewater
concentrationdata using a Bayesian hierarchical framework applied to the two
distinctdata sources. By default the model assumes a fixed generation
interval and delay from infection to the event that is counted. See the
getting started vignette for an example model specifications fitting
COVID-19 hospital admissions from a hypothetical state and wasteawter
concentration data from multiple sites within that state.
}
